% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{ancestor}
\alias{ancestor}
\title{Return index of \eqn{k}th ancestor pixel}
\usage{
ancestor(p, k)
}
\arguments{
\item{p}{A pixel index specified in nested order.}

\item{k}{A generation of an ancestor pixel.}
}
\description{
Gives the pixel at resolution \eqn{j - k} that contains \eqn{p},
where \eqn{p} is specified at resolution \eqn{j} (notice
it does not depend on \eqn{j}).
}
\examples{

 ancestor(4,2)
 ancestor(17,2)

}
