% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{area.CMBDataFrame}
\alias{area.CMBDataFrame}
\title{Area of a \code{\link{CMBDataFrame}}}
\usage{
\method{area}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame}
}
\value{
the sum of the areas of all pixels (rows) in cmbdf
}
\description{
Gives the surface on the unit sphere
that is encompassed by all pixels in \code{cmbdf}
}
