% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{area.CMBWindow}
\alias{area.CMBWindow}
\title{Get the spherical area of a \code{\link{CMBWindow}}}
\usage{
\method{area}{CMBWindow}(win)
}
\arguments{
\item{win}{a CMBWindow}
}
\value{
Tthe spherical area inside win
}
\description{
Get the spherical area of a \code{\link{CMBWindow}}
}
