% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{area}
\alias{area}
\title{Get the spherical area inside a CMBWindow}
\usage{
area(win)
}
\arguments{
\item{win}{a CMBWindow}
}
\value{
Tthe spherical area inside win
}
\description{
Get the spherical area inside a CMBWindow
}
