% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{as.CMBDataFrame}
\alias{as.CMBDataFrame}
\title{as.CMBDataFrame}
\usage{
as.CMBDataFrame(df, ordering, nside, spix)
}
\arguments{
\item{df}{Any \code{data.frame} whose rows are in HEALPix order}

\item{ordering}{character string that specifies the ordering scheme
("ring" or "nested")}

\item{nside}{an integer that specifies the Nside (resolution)
HEALPix parameter}

\item{spix}{a vector that specifies the HEALPix pixel index
corresponding to each row of \code{df}. If \code{spix} is left blank and
\code{df} is a \code{data.frame}, then \code{df} is assumed to contain data
for every pixel at resolution parameter \code{nside} (the full sky).
However, if \code{spix} is left blank and \code{df} is a \code{CMBDataFrame},
then \code{spix} is set equal to \code{pix(df)}}
}
\value{
A CMBDataFrame
}
\description{
Safely converts a \code{\link{data.frame}} to a CMBDataFrame. The
rows of the data.frame are assumed to be in the HEALPix order
given by \code{ordering}, and at the HEALPix resolution given
by \code{nside}. Coordinates, if present, are checked to correspond
to HEALPix pixel centers. The coordinates must be named either x,y,z
(cartesian) or theta, phi (spherical colatitude and longitude respectively).
}
