% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{assumedConvex<-}
\alias{assumedConvex<-}
\title{Change the \code{\link{assumedConvex}} boolean of a \code{\link{CMBWindow}}}
\usage{
assumedConvex(win, ...) <- value
}
\description{
Change the \code{\link{assumedConvex}} boolean of a \code{\link{CMBWindow}}
}
\keyword{internal}
