% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{assumedConvex}
\alias{assumedConvex}
\title{Check if a \code{\link{CMBWindow}} is assumed convex}
\usage{
assumedConvex(win, assume.convex)
}
\arguments{
\item{win}{a CMBWindow object}

\item{assume.convex}{optionally change the assumedConvex
attribute to TRUE or FALSE}
}
\description{
Check if a \code{\link{CMBWindow}} is assumed convex
}
