% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{assumedUniquePix}
\alias{assumedUniquePix}
\title{assumedUniquePix}
\usage{
assumedUniquePix(obj)
}
\arguments{
\item{obj}{Any object}
}
\value{
A boolean. This is TRUE if \code{obj}
is a CMBDataFrame or a HPDataFrame whose
rows can be assumed to correspond to unique
HEALPix pixel indices.
}
\description{
Check if an object can be assumed to have rows
that correspond to unique HEALPix pixel indices.
}
