% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnToDoDocumentation.R
\name{bar}
\alias{bar}
\title{The Bar Function}
\usage{
bar(a, b)
}
\arguments{
\item{a}{This is the first argument}

\item{b}{This is the second argument}
}
\description{
The Bar Function
}
