% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{car2sph}
\alias{car2sph}
\title{car2sph}
\usage{
car2sph(df)
}
\arguments{
\item{df}{a data.frame with columns labelled x, y and z}
}
\value{
a data.frame with columns lat and long for latitude and
longitude
}
