% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrame.R
\name{CMBDataFrame}
\alias{CMBDataFrame}
\title{CMB Data Frames}
\usage{
CMBDataFrame(CMBData = "../../CMB_map_smica1024.fits", coords = "spherical",
  includePolar = FALSE, includeMasks = FALSE, spix = NULL,
  sampleSize = NULL, Nside = NULL, ordering = NULL)
}
\arguments{
\item{CMBData}{can be the output of readFITScmb or it can be a string location of FITS file.}

\item{coords}{can be "spherical," "HEALPix" or "cartesian". If "healpix" then spix must be NULL.}

\item{includePolar}{TRUE if polarisation data is required, otherwise FALSE}

\item{includeMasks}{TRUE if TMASK and PMASK are required, otherwise FALSE}

\item{spix}{a vector of sample pixel indices}

\item{sampleSize}{if a positive integer is given a simple random sample of size sampleSize
will be taken from CMBData}

\item{Nside}{optionally specify the Nside parameter manually}

\item{ordering}{optionally specify the HEALPix ordering scheme (RING or NESTED) manually}
}
\value{
A data frame whose columns contain the pixel center coordinates (longitude, lattitude)
or (x,y,z), CMB intensities (I), and optionally polarisation (Q,U) and masks (TMASK, PMASK).
}
\description{
The function \code{CMBDataFrame} creates CMB Data Frames. These are a special type of
\link{data.frame} that carry extra information about the HEALPix ordering scheme,
coordinate system, and Nside parameter.
}
\examples{
##Method 1: Read the data while constructing the CMBDataFrame
df <- CMBDataFrame()

##Method 2: Read the data first, then construct the CMBDataFrame
cmbdat <- readFITScmb("../../CMB_map_smica1024.fits")
##We'll just take the sample pixels 1,2 and 3
df <- CMBDataFrame(CMBData = cmbdat, spix = c(1,2,3))

}
