% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBReadFITS.R
\name{CMBReadFITS2}
\alias{CMBReadFITS2}
\title{Read CMB data from a FITS file.}
\usage{
CMBReadFITS2(filename = "CMB_map_smica1024.fits", sample.size)
}
\arguments{
\item{filename}{The path to the fits file.}

\item{sample.size}{If specified then a simple random sample of the data will loaded}
}
\value{
A list containing the intensity (I), polarisation (Q, U), PMASK, TMASK and metadata.
}
\description{
\code{CMBReadFITS} is adapted from the \link{readFITS} function in package
  \href{https://cran.r-project.org/web/packages/FITSio/index.html}{FITSio}.
  \code{CMBReadFITS} is in development stage and will only work with 'CMB_map_smica1024.fits'.
  When it works, \code{CMBReadFITS} is much faster than \code{readFITS}.
  However, \code{readFITS} is more general and so is more likely to work.
}
\examples{
dat <- CMBReadFITS("CMB_map_smica1024.fits")

# View metadata
dat$header1
dat$header2
dat$resoln
dat$method
dat$coordsys
dat$nside
dat$hdr

}
