% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmbstat.R
\docType{package}
\name{cmbstat}
\alias{cmbstat}
\alias{cmbstat-package}
\title{cmbstat: A package for analysis of Cosmic Microwave Background data on a sphere}
\description{
The cmbstat package provides twelvety categories of important functions:
foo, bar, baz and others.
}
\section{Foo functions}{

The foo functions...
}

