% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindow.R
\name{CMBWindow}
\alias{CMBWindow}
\title{CMBWindow}
\usage{
CMBWindow(...)
}
\arguments{
\item{...}{arguments that must be labelled either x, y, z
(cartesian) or theta, phi (spherical, colatitude and latitude respectively).
Alternatively, a single data.frame may be passed in with columns labelled
x, y, z or theta, phi.
The rows correspond to clockwise ordered
vertices defining a spherical polygon on the surface of the unit sphere.
There must be at least 3 rows (i.e., spherical bigons are excluded).
'Clockwise' is understood from a perspective outside the sphere looking
in towards the origin.}
}
\value{

}
\description{
Add description here
}
\details{
Add further details here
}
\examples{
win <- CMBWindow(theta = c(0,1,2), phi = c(0.5, 1.2))

}
