% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{coords<-.CMBWindow}
\alias{coords<-.CMBWindow}
\title{Assign new coordinate system to CMBWindow}
\usage{
\method{coords}{CMBWindow}(x, ...) <- value
}
\description{
Assign new coordinate system to CMBWindow
}
\keyword{internal}
