% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{coords<-.data.frame}
\alias{coords<-.data.frame}
\title{Assign new coordinate system to a \code{\link{data.frame}}}
\usage{
\method{coords}{data.frame}(x, ...) <- value
}
\description{
Assign new coordinate system to a \code{\link{data.frame}}
}
\examples{

## Create df with no coords, then create df2 with cartesian coords
df <- data.frame(x = c(1,0,0), y = c(0,1,0), z = c(0,0,1))
df2 <- coords(df, new.coords = "cartesian")
df2
df

## Change the coords of df directly (to spherical)
coords(df) <- "spherical"
df

}
\seealso{
\code{\link{coords.data.frame}}
}
\keyword{internal}
