% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{coords<-.HPDataFrame}
\alias{coords<-.HPDataFrame}
\title{Assign new \code{\link{coords}} system to \code{\link{HPDataFrame}}}
\usage{
\method{coords}{HPDataFrame}(hpdf, ...) <- value
}
\description{
Assign new \code{\link{coords}} system to \code{\link{HPDataFrame}}
}
