% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{coords.CMBDataFrame}
\alias{coords.CMBDataFrame}
\title{Coordinate system from a CMBDataFrame}
\usage{
\method{coords}{CMBDataFrame}(cmbdf, new.coords)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}

\item{new.coords}{specifies the new coordinate system ("spherical" or "cartesian")
if a change of coordinate system is desired.}
}
\value{
If new.coords is unspecified, then the name of the coordinate system
of \code{cmbdf} is returned. Otherwise a new CMBDataFrame is returned
equivalent to \code{cmbdf} but having the desired change of coordinates
}
\description{
This function returns the coordinate system used in a CMBDataFrame.
The coordinate system is either "cartesian" or "spherical"
}
\details{
If a new coordinate system is specified, using e.g. new.coords = "spherical", the
coordinate system of the CMBDataFrame will be converted.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
coords(df)
coords(df, new.coords = "cartesian")

}
