% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{coords.CMBWindow}
\alias{coords.CMBWindow}
\title{Coordinate system from a \code{\link{CMBWindow}}}
\usage{
\method{coords}{CMBWindow}(win, new.coords)
}
\arguments{
\item{new.coords}{specifies the new coordinate system
("spherical" or "cartesian")
if a change of coordinate system is desired}

\item{cmbdf}{a CMBWindow.}
}
\value{
If new.coords is unspecified, then the name of the coordinate system
of \code{win} is returned. Otherwise a new CMBWindow is returned
equivalent to \code{win} but having the desired change of coordinates
}
\description{
This function returns the coordinate system used in a
\code{\link{CMBWindow}}. The coordinate system is either
"cartesian" or "spherical"
}
\details{
If a new coordinate system is specified, using e.g.
new.coords = "spherical", the
coordinate system of the CMBWindow will be converted
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
coords(df)
coords(df, new.coords = "cartesian")

}
