% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{coords.data.frame}
\alias{coords.data.frame}
\title{Change the coordinate system of a data.frame}
\usage{
\method{coords}{data.frame}(df, new.coords)
}
\arguments{
\item{df}{a data.frame with columns labelled x, y, z (for cartesian)
or theta, phi (for spherical colatitude and longitude respectively)}

\item{new.coords}{specifies the new coordinate system
("spherical" or "cartesian").}
}
\value{
A new data.frame is whose coordinates are as specified by
\code{new.coords}
}
\description{
Change the coordinate system of a data.frame
}
