% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{coords}
\alias{coords}
\title{Coordinate system from a CMBDataFrame}
\usage{
coords(cmbdf, new.coords)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}

\item{new.ordering}{specifies the new coordinate system ("spherical" or "cartesian")
if a change of coordinate system is desired.}
}
\value{
The name of the coordinate system that is used in the CMBDataFrame
}
\description{
This function returns the coordinate system used in a CMBDataFrame.
The coordinate system is either "cartesian" or "spherical"
}
\details{
If a new coordinate system is specified, using e.g. new.coords = "spherical", the
coordinate system of the CMBDataFrame will be converted.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
coords(df)
coords(df, new.coords = "cartesian")

}
