% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCMB.R
\name{covCMB}
\alias{covCMB}
\title{Covariance for CMB}
\usage{
covCMB(cmbdf, num.bins = 10, sample.size, max.dist, breaks,
  equiareal = TRUE, calc.max.dist = FALSE)
}
\arguments{
\item{cmbdf}{is a \code{\link{CMBDataFrame}} or \code{data.frame}}

\item{num.bins}{specifies the number of bins}

\item{sample.size}{optionally specify the size of a simple random
sample to take before calculating covariance. This may be useful if
the full covariance computation is too slow.}

\item{max.dist}{an optional number between 0 and pi specifying the
maximum geodesic distance between any two points in cmbdf.
For example, if cmbdf represents a full sky map
or a random sample of a full sky map then \code{max.dist = pi}.
If max.dist is known then specifying it may reduce
computation time.}

\item{breaks}{optionally specify the breaks manually using a
vector giving the break points between cells, i.e., the vector
has length \code{num.bins - 1}. If \code{equiareal = TRUE} then
these breaks should be \eqn{cos(r_i)} where \eqn{r_i} are radii,
and \eqn{cos(r_i) - cos(r_{i+1})} should be fixed for all i. If
\code{equiareal = FALSE} then these breaks should be \eqn{r_i}
and \eqn{r_{i+1} - r_i} should be fixed.}

\item{equiareal}{if TRUE then the bins have equal spherical
area. If false then the bins have equal annular widths.
Default is TRUE.}

\item{calc.max.dist}{if TRUE then the \code{max.dist} will be
calculated from the locations in cmbdf. Otherwise
either \code{max.dist} must be specified or max.dist will
default to pi.}
}
\value{
A data.frame containing sample covariance values, bin centers,
and number \code{n} of data point pairs whose distance falls in
the corresponding bin. The first
row of this data.frame corresponds to the sample variance.
If \code{breaks} are specified manually then locations of bin
centers are not returned.
}
\description{
This function provides an empirical covariance estimate for data
in a CMBDataFrame or data.frame. It places data into bins.
}
\examples{

}
