% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD-covCMB.R
\name{covCMBold}
\alias{covCMBold}
\title{(old/unworking) Covariance for CMB}
\usage{
covCMBold(df = CMBDataFrame(CMBData = "CMB_map_smica1024.fits", coords =
  "HEALPix", ordering = "nested"), rmin = 10^(-6), rmax = pi - 10^(-6),
  Nr = 10, Nside = 1024, N_x_vec = 10)
}
\arguments{
\item{rmin, rmax}{are the minimum and maximum of radii.}

\item{Nr}{is the number of radii in between rmin and rmax for which covariance of CMB is evaluated.}

\item{Nside}{is the Nside for which the HEALPix points are used.}

\item{N_x_vec}{is the number of points x for each radius, the number of points y for each
is 2^(ceil(log2((sqrt(N_x_vec))))) which is equivalent to sqrt(N_x_vec).}
}
\value{
the output is the data frame of radius r and the covariance Tcov.
}
\description{
This function computes the covariance for CMB.
It does not place data into bins, but instead generates
points on a circle of radius r and then finds the closest
HEALPix point to each, using nestSearch.
}
\examples{
# compute the covariance of CMB at Nside = 1024 and radii between 10^(-6) and pi-10^(-6) with 10 radii
covCMB(rmin = 10^(-6), rmax = pi-10^(-6), Nr = 10, Nside = 1024, N_x_vec = 10)

}
