% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{covmodelCMB}
\alias{covmodelCMB}
\title{Computes values of spherical covariance functions}
\usage{
covmodelCMB(obj, cov.model = "matern",
  cov.pars = stop("no cov.pars argument provided"), kappa = 0.5)
}
\arguments{
\item{obj}{Vector of distances between pairs of spatial locations.}

\item{cov.model}{A type of the correlation function. Available choices are: "matern",
"exponential","spherical", "powered.exponential", "cauchy", "gencauchy", "pure.nugget",
"askey", "c2wendland", "c4wendland", "sinepower", "multiquadric".}

\item{cov.pars}{A vector with two covariance parameters. The first parameter
corresponds to the variance sigma^2. The second parameter corresponds
to the range phi of the correlation function.}

\item{kappa}{A smoothness parameter of the correlation function.}
}
\value{
Values of the covariances for given distances.
}
\description{
This function computes the covariances given the separation distance of  locations.
Options for different covariance functions are available. The function uses
\code{\link[geoR]{cov.spatial}} for models from the package \strong{geoR} and
introduces additional new models.
}
\details{
See the section Details in \code{\link[geoR]{cov.spatial}}. Expressions for
the correlation functions which are not included in the packege \strong{geoR}:
}
\examples{

## Plot of the Askey covariance function

v1.f <- function(x, ...) {covmodelCMB (x, ...)}
curve( v1.f(x, cov.pars = c(1, 0.03), kappa = 3, cov.model = "askey"),
from = 0, to = 0.1, xlab = "distance", ylab = expression(cov(h)), lty = 2,
main = "covariance")

## Plot of the Askey variogram
v.f <- function(x, ...) { 1 - covmodelCMB (x, ...)}
curve(v.f(x,  cov.pars = c(1, 0.03), kappa = 3, cov.model = "askey" ),
 from = 0, to = 0.1, xlab = "distance", ylab = expression(gamma(h)), lty = 2,
 main = "variogram")

}
\references{
\strong{geoR} package, \code{\link[geoR]{cov.spatial}}

T. Gneiting. Strictly and non-strictly positive definite functions on spheres.
Bernoulli 19 (2013), no. 4, 1327-1349.
}
