% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{distBinList}
\alias{distBinList}
\title{distBinList}
\usage{
distBinList(cmbdf, breaks)
}
\arguments{
\item{cmbdf}{a data.frame whose first 3 columns represent the cartesian
coordinates x,y,z of nrow(cmbdf) points on a unit sphere.}

\item{breaks}{a vector, sorted from lowest to highest,
specifying the break points for the intervals that are
used to categorise the geodesic distances. The intervals are
open at left and closed at right.}
}
\value{
Let \eqn{x_i, x_j} be the points represented by row i and row j of
cmbdf, where\eqn{i < j}, and let \eqn{L} denote the output list
\code{L <- distBinList(cmbdf, breaks)}. Suppose the distance
\eqn{d(x_i,x_j)} falls into the \eqn{k^{th}} interval determined
by \code{breaks}, then \code{L[[i]][j-i]} will hold the value \eqn{k}.
}
\description{
Find all geodesic distances between points on the unit
sphere corresponding to the rows of the data.frame cmbdf,
then categorise these distances according to the intervals
given by the breaks argument.
}
