% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functionLearn.R
\name{foobar}
\alias{foobar}
\alias{foobar.numeric}
\alias{foobar.character}
\alias{thisIsOutOfPlace}
\title{Foo Bar generic}
\usage{
foobar(x)

\method{foobar}{numeric}(x)

\method{foobar}{character}(x)

thisIsOutOfPlace(x, y)
}
\arguments{
\item{x}{Object to foo}

\item{y}{Numeric for thisIsOutOfPlace}
}
\description{
The generic will \strong{foo} certain things, and then bar for
\emph{reasons}. Contact \email{daniel-fryer@live.com.au} for more functions
like this if you
\enumerate{
\item Have money.
\item Are willing to spend it.
}
because
\deqn{a + b = \U222B &#x222b x dx}
}
\section{Methods (by class)}{
\itemize{
\item \code{numeric}: Difference between the mean and the median.

\item \code{character}: First and last values pasted together in a string.
}}

