% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{geo2sph}
\alias{geo2sph}
\title{geo2sph}
\usage{
geo2sph(...)
}
\arguments{
\item{...}{A data.frame with columns lat and lon,
or named vectors of lat and lon.}
}
\description{
Convert latitude (lat) and longitude (lon) to spherical
coordinates (theta, phi) with theta in [0,pi] and
phi in [0,2*pi).
All values are assumed to be in radians.
}
