% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{geoArea.HPDataFrame}
\alias{geoArea.HPDataFrame}
\title{Geodesic area covered by a \code{\link{HPDataFrame}}}
\usage{
\method{geoArea}{HPDataFrame}(cmbdf)
}
\arguments{
\item{hpdf}{a HPDataFrame}
}
\value{
the sum of the areas of all pixels (rows) in hpdf
}
\description{
Gives the surface on the unit sphere
that is encompassed by all pixels in \code{hpdf}
}
