% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{geoArea}
\alias{geoArea}
\title{geoArea generic}
\usage{
geoArea(x)
}
\arguments{
\item{x}{An object.}
}
\description{
geoArea generic
}
\seealso{
\code{\link{geoArea.CMBDataFrame}}
\code{\link{geoArea.HPDataFrame}}
\code{\link{geoArea.CMBWindow}}
}
