% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathHelpers.R
\name{geoDist}
\alias{geoDist}
\title{geodesic distance on the unit sphere}
\usage{
geoDist(p1, p2)
}
\arguments{
\item{p1}{a point on the unit sphere given in cartesian coordinates (x,y,z)}

\item{p2}{a point on the unit sphere given in cartesian coordinates (x,y,z)}
}
\value{
the geodesic distance between \code{p1} and \code{p2}
}
\description{
geodesic distance on the unit sphere
}
