% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{ibp2p}
\alias{ibp2p}
\title{Computes pixel's index using its subindex within base resolution}
\usage{
ibp2p(ibp, bp, j)
}
\arguments{
\item{ibp}{The pixel index within base pixel bp, at resolution j, in nested order.}

\item{bp}{The base pixel index}

\item{j}{The resolution parameter nside = 2^j}
}
\description{
Find the pixel index \eqn{p} of a given pixel with
index \code{ibp} in base pixel \code{bp}.
}
\examples{

ibp2p(1, 1, 2)
ibp2p(1, 2, 2)

}
