% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{is.CMBDat}
\alias{is.CMBDat}
\title{Check if an object is of class CMBDat}
\usage{
is.CMBDat(cmbdf)
}
\arguments{
\item{cmbdf}{Any R object}
}
\value{
TRUE if \code{cmbdf} is a CMBDat object, otherwise FALSE
}
\description{
Check if an object is of class CMBDat
}
