% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{is.CMBDataFrame}
\alias{is.CMBDataFrame}
\title{Check if an object is of class CMBDataFrame}
\usage{
is.CMBDataFrame(cmbdf)
}
\arguments{
\item{cmbdf}{Any R object}
}
\value{
TRUE if \code{cmbdf} is a CMBDataFrame, otherwise FALSE
}
\description{
Check if an object is of class CMBDataFrame
}
