% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{is.CMBWindow}
\alias{is.CMBWindow}
\title{Check if an object is a CMBWindow}
\usage{
is.CMBWindow(win)
}
\arguments{
\item{win}{any object}
}
\value{
TRUE or FALSE depending if win is a CMBWindow
}
\description{
Check if an object is a CMBWindow
}
