% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{is.HPDataFrame}
\alias{is.HPDataFrame}
\title{Check if an object is of class \code{\link{HPDataFrame}}}
\usage{
is.HPDataFrame(hpdf)
}
\arguments{
\item{hpdf}{Any R object}
}
\value{
TRUE if \code{hpdf} is a HPDataFrame, otherwise FALSE
}
\description{
Check if an object is of class \code{\link{HPDataFrame}}
}
