% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathHelpers.R
\name{jacobiPol}
\alias{jacobiPol}
\title{Calculate Jacobi polynomial values}
\source{
\url{http://dlmf.nist.gov/18.9}
}
\usage{
jacobiPol(a, b, L, t)
}
\arguments{
\item{a, b}{The parameters of Jacobi polynomial}

\item{L}{The degree of Jacobi polynomial}

\item{t}{Given point in [-1,1].}
}
\value{
Jacobi polynomial values
}
\description{
Calculate Jacobi polynomial values of degree L at given point T in [-1,1].
}
\examples{

jacobiPol(0,0,5,0)
jacobiPol(2,-5,2,-1)
jacobiPol(1,2,4,0.5)

}
\keyword{internal}
