% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JacobiRecursive.R
\name{JacobiRecursive}
\alias{JacobiRecursive}
\title{Calculate Jacobi polynomial values of degree L at given point T in [-1,1].}
\source{
\url{http://dlmf.nist.gov/18.9}
}
\usage{
JacobiRecursive(a, b, L, T)
}
\arguments{
\item{L}{The degree of Jacobi polynomial}

\item{T}{Given point in [-1,1].}

\item{(a, b)}{The parameters of Jacobi polynomial}
}
\value{
Jacobi polynomial values
}
\description{
Calculate Jacobi polynomial values of degree L at given point T in [-1,1].
}
\examples{
JacobiRecursive(0,0,5,0)
JacobiRecursive(1,2,4,0.5)
}
\keyword{Jacobi,Orthogonal}
\keyword{polynomials.}
