% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{maxDist.CMBDataFrame}
\alias{maxDist.CMBDataFrame}
\title{Get the maximum distance between all points
in a \code{\link{CMBDataFrame}}}
\usage{
\method{maxDist}{CMBDataFrame}(x)
}
\arguments{
\item{x}{A CMBDataFrame object.}
}
\value{
Maximum distance between all points.
}
\description{
Get the maximum distance between all points
in a \code{\link{CMBDataFrame}}
}
\examples{

## For CMBDataFrame with all Healpix ponts included it must be pi
cmbdf <- CMBDataFrame(nside = 4)
pix(cmbdf)
maxDist(cmbdf)

## Example for CMBDataFrame with Healpix only ponts 1 and 2 included

cmbdf <- CMBDataFrame(nside = 4, spix =c(1,2))
pix(cmbdf)
maxDist(cmbdf)

}
