% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{maxDist.CMBWindow}
\alias{maxDist.CMBWindow}
\title{Get the maximum distance between all points
in a \code{\link{CMBWindow}}}
\usage{
\method{maxDist}{CMBWindow}(x)
}
\arguments{
\item{x}{A CMBWindow object.}
}
\value{
The maximum distance between window's points.
}
\description{
Get the maximum distance between all points
in a \code{\link{CMBWindow}}
}
\examples{

## win is a equilateral spherical triangle which sides are pi/2
win <- CMBWindow(theta = c(0,pi/2,pi/2), phi = c(0,0,pi/2))
maxDist(win)

}
