% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{maxDist}
\alias{maxDist}
\title{maxDist generic}
\usage{
maxDist(x)
}
\arguments{
\item{x}{An object.}
}
\description{
maxDist generic
}
\seealso{
\code{\link{maxDist.CMBDataFrame}}
\code{\link{maxDist.CMBWindow}}
}
