% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{maxDist_internal}
\alias{maxDist_internal}
\title{maxDist_internal}
\usage{
maxDist_internal(cmbdf)
}
\arguments{
\item{cmbdf}{a \code{data.frame} or \code{\link{CMBDataFrame}}}
}
\value{
the maximum distance between any of the
points in \code{cmbdf}
}
