% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{minDist}
\alias{minDist}
\title{minDist}
\usage{
minDist(cmbdf, point)
}
\arguments{
\item{cmbdf}{a \code{data.frame} or \code{\link{CMBDataFrame}}}

\item{point}{a point on the unit sphere in cartesian coordinates}
}
\value{
the shortest distance from \code{point} to \code{cmbdf}
}
