% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{nest2ring}
\alias{nest2ring}
\title{nest2ring}
\usage{
nest2ring(nside, pix)
}
\arguments{
\item{nside}{is the HEALPix nside parameter.}

\item{pix}{is the set or subset of pixel indices at nside.}
}
\value{
the output is the corresponding set of pixel in
the ring ordering scheme.
}
\description{
Convert from "nested" to "ring" ordering

\code{nest2ring} computes the HEALPix pixel index
in the "ring" ordering scheme from the pixel index
in the "nested" ordering scheme.
}
