% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_search.R
\name{nest_search}
\alias{nest_search}
\title{nest_search}
\usage{
nest_search(target, j2, j1 = 0, pix.j1 = 0)
}
\arguments{
\item{target}{is the target point on S^2 in spherical coordinates.}

\item{j2}{is the upper resolution.}

\item{j1}{is the lower resolution, with j1 < j2.}

\item{pix.j1}{is the initial pix index at resolution j1,
i.e., the j1-level pixel to search in. If \code{pix.j1 = 0} then
all pixels will be searched (slow).}
}
\value{
A list containing the Cartesian coordinates, \code{xyz},
and the HEALPix pixel index, \code{pix}, of the closest HEALPix
pixel center to the target point, \code{target}, at resolution j2
}
\description{
Search for the closest HEALPix pixel to a \code{target} point,
where the search is restricted to within HEALPix pixel,
\code{pix.j1}, at resolution j1. The returned value is a
HEALPix pixel (and, optionally, the cartesian coordinates of
its center) at resolution j2, where j2 > j1.
}
\details{
j1 and j2 are HEALPix resolution parameters, i.e., \eqn{Nside = 2^j}.

\code{nest_search(target, j2, j1, pix.j1)} searches within the subregion
pix.j1, where pix.j1 is a HEALPix pixel index at resolution j1.
The return value is the HEALPix point closest to \code{target}, at
resolution j2.

\code{nest_search(target, j2)} searches for the HEALPix point closest to
\code{target} at resolution j2, among all HEALPix points at resolution j2.
}
\examples{
# search for the HEALPix pixel center closest to North pole
# (0,0,1) at level 3
nest_search(target = c(0,0,1), j2 = 3, j1 = -1, plot_points=TRUE )

}
