% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestSearch.R
\name{nestSearch_step}
\alias{nestSearch_step}
\title{nestSearch_step}
\usage{
nestSearch_step(target, j1 = j2, j2, pix.j1 = 0, demo.plot = FALSE)
}
\arguments{
\item{target}{is the target point on S^2 in spherical coordinates.}

\item{j1}{is the lower resolution, with j1 < j2.}

\item{j2}{is the upper resolution.}

\item{pix.j1}{is the initial pix index at resolution j1,
i.e., the j1-level pixel to search in. If \code{pix.j1 = 0} then
all pixels will be searched (slow).}

\item{demo.plot}{If TRUE then a plot will be produced with
target pixel in yellow and closest pixel in red}
}
\value{
A list containing the Cartesian coordinates, \code{xyz},
and the HEALPix pixel index, \code{pix}, of the closest HEALPix
pixel center to the target point, \code{target}, at resolution j2
}
\description{
Search for the closest HEALPix pixel to a \code{target} point,
where the search is restricted to within HEALPix pixel,
\code{pix.j1}, at resolution j1. The returned value is a
HEALPix pixel (and, optionally, the cartesian coordinates of
its center) at resolution j2, where j2 > j1. All pixels are
assumed to be in nested ordering scheme.
}
\details{
j1 and j2 are HEALPix resolution parameters, i.e., \eqn{nside = 2^j}.

\code{nestSearch_step(target, j2, j1, pix.j1)} searches within the subregion
pix.j1, where pix.j1 is a HEALPix pixel index at resolution j1.
The return value is the HEALPix point closest to \code{target}, at
resolution j2.

Setting \code{pix.j1 = 0} (the default) searches for the HEALPix point closest
to \code{target} at resolution j2, among all HEALPix points at resolution j1.
}
\examples{
# search for the HEALPix pixel center closest to North pole
# (0,0,1) at level 3
nestSearch_step(target = c(0,0,1), j2 = 3, j1 = -1, demo.plot = TRUE )

}
