% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{nside}
\alias{nside}
\title{nside generic}
\usage{
nside(x)
}
\arguments{
\item{x}{An object.}
}
\description{
nside generic
}
\seealso{
\code{\link{nside.CMBDataFrame}}
\code{\link{nside.HPDataFrame}}
}
