% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{onBPBoundary}
\alias{onBPBoundary}
\title{a version of onBPBoundary to use with neighbours}
\usage{
onBPBoundary(se, so, j)
}
\arguments{
\item{se}{The sum of even bits, e.g. sum( f$even )}

\item{so}{The sum of odd bits, e.g. sum( f$odd )}

\item{j}{The resolution parameter nside = 2^j}
}
\description{
a version of onBPBoundary to use with neighbours
}
\keyword{internal}
