% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{ordering<-.CMBDataFrame}
\alias{ordering<-.CMBDataFrame}
\title{Assign a new ordering scheme to CMBDataFrame}
\usage{
\method{ordering}{CMBDataFrame}(cmbdf, ...) <- value
}
\description{
Assign a new ordering scheme to CMBDataFrame
}
\examples{
cmbdf <- CMBDataFrame(n = 1, ordering = "ring")
ordering(cmbdf) <- "nested"
ordering(cmbdf)

}
\seealso{
\code{\link{ordering.CMBDataFrame}}
}
\keyword{internal}
