% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{ordering<-.HPDataFrame}
\alias{ordering<-.HPDataFrame}
\title{Assign new ordering scheme to HPDataFrame}
\usage{
\method{ordering}{HPDataFrame}(x, ...) <- value
}
\description{
Assign new ordering scheme to HPDataFrame
}
\keyword{internal}
