% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{ordering<-}
\alias{ordering<-}
\title{ordering generic}
\usage{
ordering(x, ...) <- value
}
\description{
ordering generic
}
\seealso{
\code{\link{ordering.CMBDataFrame}}
\code{\link{ordering.HPDataFrame}}
}
\keyword{internal}
