% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{ordering.CMBDataFrame}
\alias{ordering.CMBDataFrame}
\title{HEALPix ordering scheme from a CMBDataFrame}
\usage{
\method{ordering}{CMBDataFrame}(cmbdf, new.ordering)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}

\item{new.ordering}{specifies the new ordering ("ring" or "nest") if a change of ordering
scheme is desired.}
}
\value{
The name of the HEALPix ordering scheme that is used in the CMBDataFrame cmbdf
}
\description{
This function returns the HEALPix ordering scheme from a CMBDataFrame.
The ordering scheme is either "ring" or "nested".
}
\details{
If a new ordering is specified, using e.g. new.ordering = "ring", the
ordering scheme of the CMBDataFrame will be converted.
}
\examples{
df <- CMBDataFrame(nside = 1, ordering = "nested")
ordering(df)
df1 <- ordering(df, new.ordering = "ring")
ordering(df1)

}
