% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{ordering.HPDataFrame}
\alias{ordering.HPDataFrame}
\title{HEALPix ordering scheme from a HPDataFrame}
\usage{
\method{ordering}{HPDataFrame}(hpdf, new.ordering)
}
\arguments{
\item{hpdf}{a \code{\link{HPDataFrame}}.}

\item{new.ordering}{specifies the new ordering ("ring" or "nest")
if a change of ordering scheme is desired.}
}
\value{
The name of the HEALPix ordering scheme that is used in the
HPDataFrame hpdf, or a new hpdf with the desired new.ordering
}
\description{
This function returns the HEALPix ordering scheme from a HPDataFrame.
The ordering scheme is either "ring" or "nested". If a new ordering
is specified, using e.g. \code{new.ordering = "ring"}, the
ordering scheme of the HPDataFrame will be converted.
}
\examples{
## Plot using indices
df <- HPDataFrame(I = rep(0,12), nside = 1, ordering = "nested")
ordering(df)
ordering(df, new.ordering = "ring")

## Plot using coordinates
hp1 <- HPDataFrame(x = c(1,0,0),
                   y = c(0,1,0),
                   z = c(0,0,1),
                   nside = 1,
                   auto.spix = TRUE)
plot(hp, size = 5, hp.boundaries = 1)

}
