% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{p2bp}
\alias{p2bp}
\title{p2bp}
\usage{
p2bp(p, j)
}
\arguments{
\item{p}{The pixel index at resolution j, in nested order.}

\item{j}{The resolution parameter nside = 2^j}
}
\description{
The base pixel to which pixel p belongs at resolution j
}
