% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{p2ibp}
\alias{p2ibp}
\title{p2ibp}
\usage{
p2ibp(p, j)
}
\arguments{
\item{p}{The pixel index at resolution j, in nested order.}

\item{j}{The resolution parameter nside = 2^j}
}
\description{
Convert a pixel index p to its index within
the base pixel to which p belongs
}
