% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/NestSearch.R
=======
% Please edit documentation in R/nestSearch.R
>>>>>>> 35b1876d6781bb63e95d42038d82f7f11623d604
\name{parent}
\alias{parent}
\title{parent}
\usage{
parent(p)
}
\arguments{
\item{p}{A pixel index specified in nested order.}
}
\description{
Gives the pixel at resolution j - 1 that contains p,
where p is specified at resoution j (notice it does not depend on j).
}
