% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{pix<-.CMBDataFrame}
\alias{pix<-.CMBDataFrame}
\title{Assign new pixel indices to a CMBDataFrame}
\usage{
\method{pix}{CMBDataFrame}(x, ...) <- value
}
\description{
Assign new pixel indices to a CMBDataFrame
}
\keyword{internal}
