% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{pix<-}
\alias{pix<-}
\title{pix generic}
\usage{
pix(x, ...) <- value
}
\description{
Detailed descriptions and and examples can be found in documentation for specific
pix functions \code{\link{pix.CMBDataFrame}}, \code{\link{pix.HPDataFrame}}
}
\seealso{
\code{\link{pix.CMBDataFrame}}
\code{\link{pix.HPDataFrame}}
}
\keyword{internal}
