% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{pix.CMBDataFrame}
\alias{pix.CMBDataFrame}
\title{HEALPix pixel indices from \code{\link{CMBDataFrame}}}
\usage{
\method{pix}{CMBDataFrame}(cmbdf, new.pix)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}

\item{new.pix}{optional vector of pixel indices}
}
\value{
The vector of HEALPix pixel indices or, if new.pix is specified,
a new CMBDataFrame.
}
\description{
If new.pix is unspecified then this function returns the vector of
HEALPix pixel indices from a CMBDataFrame. If new.pix is specified then
this function returns a new CMBDataFrame with pixel indices new.pix
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
pix(df)

df.new <- pix(df, new.pix= 1:10)
pix(df.new)

}
