% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{pix}
\alias{pix}
\title{HEALPix pixel indices from CMBDataFrame}
\usage{
pix(cmbdf)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}
}
\value{
The vector of HEALPix pixel indices
}
\description{
This function returns the vector of HEALPix pixel indices from a CMBDataFrame
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
pix(df)

}
