% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pix2ang.R
\name{pix2ang}
\alias{pix2ang}
\title{Pix to Angle}
\usage{
pix2ang(nSide, order_ring, Pix)
}
\arguments{
\item{nSide}{is the Nside of the HEALPix points.}

\item{order_ring}{TRUE if the HEALPix order is in the ring order; 
FALSE if in the nested order.}

\item{Pix}{is the set of Pix index.}
}
\value{
the output is a 2 by length(Pix) matrix of spherical coordinates (theta,phi)' of HEALPix points
           at nSide at Pix index in the order ord.
}
\description{
\code{pix2ang} computes the Cartesian cooridinates of the HEALPix points 
   at Pix index at nSide in order ord.
}
\examples{
# if nargs() < 3, Pix = 1:nPix, nPix = 12*nSide^2. That is, taking all
   all HEALPix in the ring order.
Nside <- 8
pix2ang(Nside,TRUE)

# comput the HEALPix points in spherical coordinates at Nside in the nest order at pix indices from Pix
Nside <- 8
Pix <- c(1,2,23)
pix2ang(Nside,FALSE,Pix)

}
