% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pix2angC}
\alias{pix2angC}
\title{pix2angC}
\arguments{
\item{Nside}{The number of cuts to a HEALPix base resolution pixel.}

\item{Nest}{A boolean, TRUE if ordering scheme is NESTED and FALSE otherwise.}
}
\description{
Converts HEALPix pixel scheme to spherical coordinates.
}
\details{
This is a place holder
}
