% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{pix2coords}
\alias{pix2coords}
\title{pix2coords}
\usage{
pix2coords(nside, coords = "cartesian", ordering = "nested", spix)
}
\arguments{
\item{nside}{the nside parameter}

\item{coords}{'cartesian' or 'spherical' coordinates}

\item{ordering}{'ring' or 'nested' ordering}

\item{spix}{optional integer or vector of sample pixel indices}
}
\value{
a data.frame with columns 'x', 'y', 'z' (cartesian) or
'theta', 'phi' (spherical)
}
\description{
convert HEALPix pixel indices to cartesian or spherical coordinates
}
