% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OLD-pix2vec.R
\name{pix2vec}
\alias{pix2vec}
\title{pix2vec}
\usage{
pix2vec(Nside = 16, order = "ring", spix)
}
\arguments{
\item{Nside}{is the Nside of the HEALPix points.}

\item{order}{specifies the ordering scheme used for the input
pixels \code{spix}. If \code{order = "nested"} then the input
pixels are converted to ring ordering. If \code{order = "ring"}
then no conversion is done.}

\item{spix}{is a vector of one or more pixel indices whose ordering
scheme is given by \code{order} and whose Nside parameter is given
by \code{Nside}. If spix = 0 then all pixel indices at
\code{Nside} will be used.}
}
\value{
Output is a data.frame with \code{length(spix)} rows, or
\eqn{12*Nside^2} rows if \code{spix = 0},
and 3 columns of Cartesian coordinates (x,y,z) of HEALPix points
at \code{Nside} in ring ordering scheme.
}
\description{
Compute the Cartesian cooridinates of the HEALPix points
at indices \code{spix} at \code{Nside} in the specified
ordering scheme.
}
\examples{
# Taking all HEALPix points at Nside = 8, using ring order.
pix2vec(Nside = 8, order = "ring")

# Compute the HEALPix points in Cartesian coordinates at Nside = 8,
# in nested ordering scheme, at the indices from spix
Pix <- c(1,2,23)
pix2vec(Nside = 8, order = "nested", spix = Pix)

}
