% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{pixelArea}
\alias{pixelArea}
\title{pixelArea}
\usage{
pixelArea(cmbdf)
}
\arguments{
\item{cmbdf}{a \code{\link{CMBDataFrame}}}
}
\value{
the area of a single HEALPix pixel
at the \code{nside} resolution of \code{cmbdf}
}
\description{
Get the area of a single HEALPix pixel
}
