% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestSearch.R
\name{pixelWindow}
\alias{pixelWindow}
\title{Pixel window}
\usage{
pixelWindow(j1, j2, pix.j1)
}
\arguments{
\item{j1}{is the lower resolution, with j1 < j2}

\item{j2}{the upper resolution}

\item{pix.j1}{the pixel index at resolution j1 within which
all pixels from resolution j2 will be returned. \code{pix.j1} can
also be a vector of non-zero pixel indices.}
}
\value{
All pixels in resolution j2 that fall within the pixel
pix.j1 specified at resolution j1
}
\description{
All pixels are assumed to be in nested ordering
}
