% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{plot.CMBcorrelation}
\alias{plot.CMBcorrelation}
\title{Plot CMBcorrelation}
\usage{
\method{plot}{CMBcorrelation}(x, ...)
}
\arguments{
\item{x}{An object of class CMBCorrelation.}

\item{...}{Extra arguments passed to plot.default.}
}
\description{
Plot CMBcorrelation
}
