% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{plot.CMBcovariance}
\alias{plot.CMBcovariance}
\title{Plot CMBcovariance}
\usage{
\method{plot}{CMBcovariance}(x, ...)
}
\arguments{
\item{x}{An object of class CMBCovariance.}

\item{...}{Extra arguments passed to plot.default.}
}
\description{
Plot CMBcovariance
}
