% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{plot.CMBvariogram}
\alias{plot.CMBvariogram}
\title{Plot CMBvariogram, CMBcovariance, CMBcorrelation}
\usage{
\method{plot}{CMBvariogram}(x)
}
\description{
Plot CMBvariogram, CMBcovariance, CMBcorrelation
}
