% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{plot.CMBWindow}
\alias{plot.CMBWindow}
\title{visualise a \code{\link{CMBWindow}}}
\usage{
\method{plot}{CMBWindow}(win, add = TRUE, type = "l", col = "red",
  size = 2, box = FALSE, axes = FALSE, aspect = FALSE, back.col, ...)
}
\arguments{
\item{win}{a CMBWindow}

\item{add}{if TRUE then this plot will be added to any existing plot.
Note that if \code{back.col} (see below) is specified then a new plot
window will be opened and \code{add = TRUE} will have no effect}

\item{type}{a single character indicating the type of item to plot.
Supported types are: 'p' for points, 's' for spheres, 'l' for lines,
'h' for line segments from z = 0, and 'n' for nothing.}

\item{col}{specify the colour(s) of the plotted points}

\item{size}{the size of plotted points}

\item{box}{whether to draw a box}

\item{axes}{whether to draw axes}

\item{aspect}{either a logical indicating whether to adjust the
aspect ratio, or a new ratio.}

\item{back.col}{specifies the background colour of
the plot. This argument is passed to rgl::bg3d.}

\item{...}{arguments passed to rgl::plot3d}

\item{eps}{the geodesic distance between consecutive points to draw
on the window boundary}
}
\description{
visualise a \code{\link{CMBWindow}}
}
