% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{plot}
\alias{plot}
\title{plot generic}
\usage{
plot(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Extra arguments.}
}
\description{
Plot empirical variogram type functions. Detailed descriptions and examples can be found in documentation for specific
plot functions \code{\link{plot.variogram}}, \code{\link{plot.CMBcovariance}},
\code{\link{plot.CMBcorrelation}}
}
\seealso{
\code{\link{plot.variogram}}
\code{\link{plot.CMBcovariance}}
\code{\link{plot.CMBcorrelation}}
}
