% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCMB.R
\name{plotCMB}
\alias{plotCMB}
\title{Plot CMB Data}
\usage{
plotCMB(CMBDataFrame)
}
\arguments{
\item{CMBDataFrame}{a CMB Data Frame with either spherical or cartesian coordinates.}
}
\value{
A plot of the CMB data.
}
\description{
This function produces a plot from a CMB Data Frame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sampleSize = 800000)
plotCMB(df)

}
