% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInConvexPolygon}
\alias{pointInConvexPolygon}
\title{pointInConvexPolygon}
\usage{
pointInConvexPolygon(df, win)
}
\arguments{
\item{df}{a data.frame with columns x, y, z for cartesian coordinates.
The rows represent points on the surface of a unit sphere}

\item{win}{a data.frame with columns x, y, z for cartesian coordinates.
The rows represent clockwise oriented vertices of a convex spherical
polygon that lies entirely within one open hemisphere of the unit sphere.}
}
\value{
a logical vector indicated which rows of \code{df}
lie within the spherical convex polygon determined by \code{win}
}
