% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInConvexPolygonHP}
\alias{pointInConvexPolygonHP}
\title{pointInConvexPolygonHP}
\usage{
pointInConvexPolygonHP(nside, nested, win, spix = NULL)
}
\arguments{
\item{nside}{the nside parameter at which to find pixels}

\item{nested}{Set to TRUE for NESTED ordering scheme and FALSE for RING}

\item{win}{a data.frame with columns x, y, z for cartesian coordinates
The rows represent clockwise oriented vertices of a convex spherical
polygon that lies entirely within one open hemisphere of the unit sphere}

\item{spix}{Optional integer or vector of sample pixel indices. If \code{spix}
is unspecified then all pixels at \code{nside} are used}
}
\value{
a logical vector indicated which pixels in \code{spix}
lie within the spherical convex polygon determined by \code{win}
}
