% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInDisc}
\alias{pointInDisc}
\title{pointInDisc}
\usage{
pointInDisc(df, win)
}
\arguments{
\item{df}{a data.frame with columns x, y, z for cartesian coordinates.
The rows represent points on the surface of a unit sphere}

\item{win}{a data.frame with columns x, y, z for the cartesian coordinates
of a point on the unit sphere, representing a disc center, and column r for
the radius or that disc.}
}
\value{
a logical vector indicated which rows of \code{df}
lie within the spherical disc determined by \code{win}
}
