% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInDiscHP}
\alias{pointInDiscHP}
\title{pointInDiscHP}
\usage{
pointInDiscHP(nside, nested, win, spix = NULL)
}
\arguments{
\item{nside}{the nside parameter at which to find pixels}

\item{nested}{Set to TRUE for NESTED ordering scheme and FALSE for RING}

\item{win}{a data.frame with columns x, y, z for the cartesian coordinates
of a point on the unit sphere, representing a disc center, and column r for
the radius or that disc}

\item{spix}{Optional integer or vector of sample pixel indices. If \code{spix}
is unspecified then all pixels at \code{nside} are used}
}
\value{
a logical vector indicated which pixels in \code{spix}
lie within the spherical disc determined by \code{win}
}
