% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pointInPolygon}
\alias{pointInPolygon}
\title{pointInPolygon}
\usage{
pointInPolygon(df, win)
}
\arguments{
\item{df}{a data.frame with columns x, y, z for cartesian coordinates.
The rows represent points on the surface of a unit sphere}

\item{win}{a data.frame with columns x, y, z for cartesian coordinates.
The rows represent clockwise oriented vertices of a spherical polygon.}
}
\value{
a logical vector indicated which rows of \code{df}
lie within the spherical polygon determined by \code{win}
}
