% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{print.CMBDataFrame}
\alias{print.CMBDataFrame}
\title{Print CMBDataFrame}
\usage{
\method{print}{CMBDataFrame}(x, ...)
}
\arguments{
\item{x}{A \code{\link{CMBDataFrame}}.}

\item{...}{arguments passed to \code{\link[tibble:print.tbl_df]{print.tbl_df}}}
}
\value{
Prints contents of the CMB data frame to the console.
}
\description{
This function neatly prints the contents of a CMBDataFrame.
}
\examples{
## First download the map
# downloadCMBMap(foreground = "smica", nside = 1024)
# df <- CMBDataFrame("CMB_map_smica1024.fits")
# print(df)

}
