% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{print.summary.CMBDataFrame}
\alias{print.summary.CMBDataFrame}
\title{Print a summary of a CMBDataFrame}
\usage{
\method{print}{summary.CMBDataFrame}(x, ...)
}
\arguments{
\item{x}{a \code{summary.CMBDataFrame} object, i.e.,
the output of \code{\link{summary.CMBDataFrame}}}
}
\description{
Print a summary of a CMBDataFrame
}
\keyword{internal}
