% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{print.summary.CMBWindow}
\alias{print.summary.CMBWindow}
\title{Print a summary of a \code{\link{CMBWindow}}}
\usage{
\method{print}{summary.CMBWindow}(x, ...)
}
\arguments{
\item{x}{a \code{summary.CMBWindow} object, i.e.,
the output of \code{\link{summary.CMBWindow}}}
}
\description{
Print a summary of a \code{\link{CMBWindow}}
}
\keyword{internal}
