% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{print.summary.HPDataFrame}
\alias{print.summary.HPDataFrame}
\title{Print a summary of a HPDataFrame}
\usage{
\method{print}{summary.HPDataFrame}(x, ...)
}
\arguments{
\item{x}{a \code{summary.HPDataFrame} object, i.e.,
the output of \code{\link{summary.HPDataFrame}}}
}
\description{
Print a summary of a HPDataFrame
}
\keyword{internal}
