% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFITScmb.R
\name{readFITScmb}
\alias{readFITScmb}
\title{Read CMB data from a FITS file.}
\usage{
readFITScmb(filename = "CMB_map_smica1024.fits")
}
\arguments{
\item{filename}{The path to the fits file.}
}
\value{
A list containing the intensity (I), polarisation (Q, U), PMASK, TMASK and metadata.
}
\description{
\code{readFITScmb} is adapted from the \link{readFITS} function in package
  \href{https://cran.r-project.org/web/packages/FITSio/index.html}{FITSio}.
  \code{readFITScmb} is in development stage and will only work with 'CMB_map_smica1024.fits'.
  When it works, \code{readFITScmb} is much faster than \code{readFITS}.
  However, \code{readFITS} is more general and so is more likely to work.
}
\examples{
  readFITScmb("CMB_map_smica1024.fits")

}
