% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{resolution}
\alias{resolution}
\title{Get the arcmin resolution from a \code{\link{CMBDataFrame}}}
\usage{
resolution(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}
}
\value{
The arcmin resolution as specified by the FITS file where the
data was sourced
}
\description{
Get the arcmin resolution from a \code{\link{CMBDataFrame}}
}
