% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{sampleCMB}
\alias{sampleCMB}
\title{Take a simple random sample from a CMBDataFrame}
\usage{
sampleCMB(cmbdf, sample.size)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}

\item{sample.size}{the desired sample size.}
}
\value{
A CMBDataFrame with size sample.size,
whose rows comprise a simple random sample of the rows
from the input CMBDataFrame.
}
\description{
This function returns a CMBDataFrame with size sample.size,
whose rows comprise a simple random sample of the rows
from the input CMBDataFrame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits")
plot(sampleCMB(df, sample.size = 800000))

}
