% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{separatingNside}
\alias{separatingNside}
\title{separatingNside}
\usage{
separatingNside(df)
}
\arguments{
\item{df}{A data.frame. Must have columns x,y,z
for Cartesian coordinates that must represent points
on the unit sphere.}
}
\value{
An nside (power of 2) such that all points
in \code{df} will belong to unique pixels.
}
\description{
separatingNside
}
\examples{
sky <- CMBDataFrame(nside = 16, coords = "cartesian", ordering = "nested")
separatingNside(sky)

}
\keyword{internal}
