% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/NestSearch.R
=======
% Please edit documentation in R/nestSearch.R
>>>>>>> 35b1876d6781bb63e95d42038d82f7f11623d604
\name{siblings}
\alias{siblings}
\title{siblings}
\usage{
siblings(p)
}
\arguments{
\item{p}{Pixel index in nested order.}
}
\description{
The siblings of pixel p are defined as the
children of the parent of p. Note this is resolution independent.
}
