% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MathHelpers.R
\name{sph2car}
\alias{sph2car}
\title{Convert from spherical to cartesian}
\usage{
sph2car(p)
}
\arguments{
\item{p}{a point in latitude and longitude \code{c(lat, long)} or
a data.frame containing columns lat and long}
}
\value{
the cartesian coordinates of p
}
\description{
Convert from spherical to cartesian
}
