% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalHarmonics.R
\name{SphericalHarmonics}
\alias{SphericalHarmonics}
\title{Compute spherical harmonic values at given points on the sphere.}
\usage{
SphericalHarmonics(L, m, xyz)
}
\arguments{
\item{L}{The degree of spherical harmonic}

\item{m}{The order number of the degree-L spherical harmonic}

\item{xyz}{Given points in 3D cartesian coordinates}
}
\value{
The spherical harmonic values
}
\description{
The function \code{SphericalHarmonics} computes the spherical harmonic values on the given 3D cartesian coordinates.
}
\examples{
SphericalHarmonics(5,2,c(0,1,0))
SphericalHarmonics(5,2,diag(3))
}
\references{
Hesse, K., Sloan, I. H., & Womersley, R. S. (2010). 
Numerical integration on the sphere. In Handbook of Geomathematics (pp. 1185-1219). 
Springer Berlin Heidelberg.
}
\keyword{harmonic}
\keyword{spherical}
