% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnToDoDocumentation.R
\name{sum}
\alias{sum}
\alias{plus}
\alias{summation}
\title{Sum of vector elements.}
\usage{
sum(..., na.rm = TRUE)
}
\arguments{
\item{...}{Numeric, complex or logical vectors.}

\item{na.rm}{A logical scalar. Should missing values (including NaN) be
removed?}
}
\value{
If all inputs are integer and logical, then the output will be an
  integer. If integer overflow
  \url{http://en.wikipedia.org.wiki/Integer_overflow} occurs, the output will
  be NA with a warning. Otherwise it will be length-one numeric or complex
  vector.

  Zero-length vectors have sum 0 by definition.
}
\description{
\code{sum} returns the sum of all the values present in its arguments
}
\details{
This is a generic function: methods can be defined for it directly or via the
\code{\link{Summary}} group generic. For this to work properly, the arguments
\code{...} should be unnamed, and dispatch is on the first argument.
}
\section{Practice Section}{
 Just adding this text to see what roxygen does.
}

\examples{
sum(1:10)
sum(1:5, 6:10)
sum(F,F,F,T,T)
\dontrun{sum("a")}
}
\seealso{
Look at this link in my package: \code{\link{add}}

  and this link in the \code{\link{stats}} package: \code{\link[stats]{var}}

  and my cool URL \url{www.sciencestems.com}.
}
\keyword{methods}
\keyword{misc}
