% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{summary.CMBDataFrame}
\alias{summary.CMBDataFrame}
\title{Summarise a \code{\link{CMBDataFrame}}}
\usage{
\method{summary}{CMBDataFrame}(cmbdf, intensities = "I")
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}

\item{intensities}{the name of a column specifying CMB intensities
(or potentially another numeric quantity of interest)}
}
\value{
A summary
}
\description{
This function produces a summary from a CMBDataFrame.
}
