% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{summary.CMBWindow}
\alias{summary.CMBWindow}
\title{Summarise a \code{\link{CMBWindow}}}
\usage{
\method{summary}{CMBWindow}(win)
}
\arguments{
\item{cmbdf}{a CMBWindow}
}
\value{
A summary
}
\description{
This function produces a summary from a CMBWindow
}
