% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{summary.HPDataFrame}
\alias{summary.HPDataFrame}
\title{Summarise a \code{\link{HPDataFrame}}}
\usage{
\method{summary}{HPDataFrame}(hpdf, intensities = "I")
}
\arguments{
\item{hpdf}{a HPDataFrame.}

\item{intensities}{the name of a column specifying intensities
(or potentially another numeric quantity of interest)}
}
\value{
A summary includes window's type and area,
total area covered by observations,
and main statistcs for intensity values
}
\description{
This function produces a summary from a HPDataFrame.
}
