% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/developerHelpers.R
\name{tabular}
\alias{tabular}
\title{Tabular Function For rcosmo Developers}
\usage{
tabular(df, ...)
}
\arguments{
\item{df}{A data frame}

\item{...}{Arguments passed to \code{\link{format}}}
}
\description{
Turns an R data frame into the correct format for a roxygen2 table.
}
\examples{
cat(tabular(mtcars[1:5,1:5]))
}
\seealso{
Courtesy of Hadley Wickham, \href{http://r-pkgs.had.co.nz/man.html}{here}.
}
