% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{window<-.CMBDataFrame}
\alias{window<-.CMBDataFrame}
\title{Assign a new \code{\link{CMBWindow}} to a \code{\link{CMBDataFrame}}}
\usage{
\method{window}{CMBDataFrame}(x, ...) <- value
}
\description{
Assign a new \code{\link{CMBWindow}} to a \code{\link{CMBDataFrame}}
}
\keyword{internal}
