% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{window<-}
\alias{window<-}
\title{Assign a new \code{\link{CMBWindow}} to a \code{\link{CMBDataFrame}}}
\usage{
window(cmbdf, ...) <- value
}
\description{
Assign a new \code{\link{CMBWindow}} to a \code{\link{CMBDataFrame}}
}
