% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{winType<-}
\alias{winType<-}
\title{Assign new \code{\link{winType}} to a \code{\link{CMBWindow}}}
\usage{
winType(win, ...) <- value
}
\description{
Assign new \code{\link{winType}} to a \code{\link{CMBWindow}}
}
\seealso{
\code{\link{winType}}
}
