% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{as.CMBDataFrame}
\alias{as.CMBDataFrame}
\title{as.CMBDataFrame}
\usage{
as.CMBDataFrame(df, coords, ordering, nside)
}
\arguments{
\item{df}{Any data.frame with a column labelled "I" for intensities}

\item{coords}{specifies the coordinate system to be "spherical",
"cartesian" or unspecified (HEALPix only). If "spherical" then df
must have columns named "theta" and "phi" (colatitude and longitude
respectively). If "cartesian" then df
must have columns named "x", "y", and "z"}

\item{ordering}{specifies the ordering scheme ("ring" or "nested")}

\item{nside}{specifies the Nside parameter}
}
\value{
A CMBDataFrame
}
\description{
Safely converts a \code{\link{data.frame}} to a CMBDataFrame
}
