% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/NestSearch.R
=======
% Please edit documentation in R/nestSearch.R
>>>>>>> 35b1876d6781bb63e95d42038d82f7f11623d604
\name{borderPattern}
\alias{borderPattern}
\title{Border pattern is resolution independent}
\usage{
borderPattern(ptype)
}
\arguments{
\item{pype}{is the output of onBPBoundary}
}
\value{
the output is useful as an index
to the output of baseNeighbours. It will
return the correct neighbouring base
pixel in each direction. The
0 indicates to stay in current BP.
}
\description{
Border pattern is resolution independent
}
\keyword{internal}
