% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{car2sph}
\alias{car2sph}
\title{car2sph}
\usage{
car2sph(df)
}
\arguments{
\item{df}{a data.frame with columns labelled x, y and z}
}
\value{
a data.frame with columns theta and phi for colatitude and
longitude in ranges \eqn{[0,pi]} and \eqn{[0,2pi]} respectively
}
