% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{children}
\alias{children}
\title{children}
\usage{
children(p)
}
\arguments{
\item{p}{A pixel index specified in nested order.}
}
\description{
Gives the pixels at resolution j + 1 that are contained in p,
where p is specified at resoution j (notice it does not depend on j).
}
