% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{children}
\alias{children}
\title{Return children of a pixel}
\usage{
children(p)
}
\arguments{
\item{p}{A pixel index specified in nested order.}
}
\description{
Gives four pixels at resolution \eqn{j + 1} that are contained in \eqn{p},
where \eqn{p}is a pixel specified at resolution \eqn{j} (notice it does not
depend on \eqn{j}).
}
\examples{
children(11)

}
