% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrame.R
\name{CMBDat}
\alias{CMBDat}
\title{CMBDat class}
\usage{
CMBDat()
}
\description{
The function \code{CMBReadFITS} creates objects of class \code{CMBDat}.
These are lists containing header information and other metadata as well
as an element called data, whose columns may include, for example, the
intensity (I), polarisation (Q, U), PMASK and TMASK. Also it can be
\code{\link{mmap}} object that points to the CMB map data table in the FITS file.
}
\examples{
cmbdat <- CMBReadFITS("CMB_map_smica1024.fits", mmap = TRUE)
class(cmbdat)
str(cmbdat)
}
