% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDat.R
\name{CMBDat}
\alias{CMBDat}
\title{Read CMB data from a FITS file.}
\usage{
CMBDat(filename, mmap = FALSE, spix)
}
\arguments{
\item{filename}{The path to the fits file.}

\item{mmap}{A boolean indicating whether to use memory mapping.}

\item{spix}{The sample pixels (rows) to read from the FITS file
<<<<<<< HEAD
=======
<<<<<<< HEAD
=======
<<<<<<< HEAD
=======
<<<<<<< HEAD
binary data table (optional)}
=======
>>>>>>> 44dd89e8e668f9f0b76d46701c5fe117e7153762
binary data table (optional).}

\item{filename}{The path to the fits file.}

\item{mmap}{A boolean indicating whether to use memory mapping.}

\item{spix}{The sample pixels (rows) to read from the FITS file
>>>>>>> bb9ad92b0afd39d554f3654f0e40147480993019
>>>>>>> ac91aaff00269ebfbdaa90810d10e9a0738d25bb
binary data table (optional)}
}
\value{
A list containing header information and other metadata
as well as an element called \code{data} where:
If \code{mmap = FALSE} then a \code{data.frame} is
included, named \code{data}, whose columns may include, for
example, the intensity (I), polarisation (Q, U), PMASK and TMASK.
If \code{mmap = TRUE} then a \code{\link{mmap}} object is returned
that points to the CMB map data table in the FITS file.
}
\description{
\code{CMBDat} is adapted from the \code{\link{readFITS}}
function in package
  \href{https://cran.r-project.org/web/packages/FITSio/index.html}{FITSio}.
  When it works, \code{CMBDat} is much faster than
  \code{\link{readFITS}}.
  However, \code{\link{readFITS}} is more general, i.e.,
  it applies to FITS files that do not contain CMB data.
}
\details{
The function \code{CMBDat} creates objects of class \code{CMBDat}.
These are lists containing header information and other metadata as well
as an element called data, whose columns may include, for example, the
intensity (I), polarisation (Q, U), PMASK and TMASK. It also may contain an
\code{\link{mmap}} object that points to the CMB map data table in the FITS
file.
}
\examples{
cmbdat <- CMBDat("CMB_map_smica1024.fits", mmap = TRUE)
class(cmbdat)
str(cmbdat)

# View metadata
dat$header1
dat$header2
dat$resoln
dat$method
dat$coordsys
dat$nside
dat$hdr

}
