% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBReadFITS.R
\name{CMBReadFITS}
\alias{CMBReadFITS}
\title{Read CMB data from a FITS file.}
\usage{
CMBReadFITS(filename = "CMB_map_smica1024.fits", sample.size)
}
\arguments{
\item{filename}{The path to the fits file.}
}
\value{
A list containing a matrix named \code{data} whose columns may include
the intensity (I), polarisation (Q, U), PMASKa and TMASK. Other list
elements correspond to header information and other metadata.
}
\description{
\code{CMBReadFITS} is adapted from the \link{readFITS} function in package
  \href{https://cran.r-project.org/web/packages/FITSio/index.html}{FITSio}.
  \code{CMBReadFITS} is in development stage and will only work with 'CMB_map_smica1024.fits'.
  When it works, \code{CMBReadFITS} is much faster than \code{readFITS}.
  However, \code{readFITS} is more general and so is more likely to work.
}
\examples{
dat <- CMBReadFITS("CMB_map_smica1024.fits")

# View metadata
dat$header1
dat$header2
dat$resoln
dat$method
dat$coordsys
dat$nside
dat$hdr

}
