% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindow.R
\name{CMBWindow}
\alias{CMBWindow}
\title{CMBWindow}
\usage{
CMBWindow(..., r, set.minus = FALSE)
}
\arguments{
\item{...}{these arguments are compulsory and must be labelled either x, y, z
(cartesian) or theta, phi (spherical, colatitude and longitude respectively).
Alternatively, a single data.frame may be passed in with columns labelled
x, y, z or theta, phi.}

\item{r}{if a disc type window is required then this specifies the
radius of the disc}

\item{set.minus}{when \code{TRUE} the window will be the unit
sphere minus the window specified}
}
\value{

}
\description{
Create a CMBWindow: Either a polygon or a disc type
}
\details{
If \code{r} is unspecified then the rows of \code{...} correspond to
counter-clockwise ordered vertices defining a spherical polygon
on the unit sphere.
In this case, there must be at least 3 rows (vertices).
On the other hand,
if \code{r} is specified then \code{...} must have just one row, and this
row is taken to be the center of a disc of radius \code{r}
}
\examples{
win <- CMBWindow(theta = c(0,pi/2,pi/2), phi = c(0,0,pi/2))

}
