% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{coords<-.CMBDataFrame}
\alias{coords<-.CMBDataFrame}
\title{Assign new \code{\link{coords}} system to \code{\link{CMBDataFrame}}}
\usage{
\method{coords}{CMBDataFrame}(cmbdf, ...) <- value
}
\description{
Assign new \code{\link{coords}} system to \code{\link{CMBDataFrame}}
}
