% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{coords<-.CMBDataFrame}
\alias{coords<-.CMBDataFrame}
\title{Assign new coordinate system to a \code{\link{CMBDataFrame}}}
\usage{
\method{coords}{CMBDataFrame}(x, ...) <- value
}
\description{
Assign new coordinate system to a \code{\link{CMBDataFrame}}
}
\examples{

## Create df with no coords, then create df2 with cartesian coords
df <- CMBDataFrame(nside = 16)
df
coords(df)
df2 <- coords(df, new.coords = "cartesian")
coords(df2)
coords(df)

## Change the coords of df directly (to spherical)
coords(df) <- "spherical"
df

}
\seealso{
\code{\link{coords.CMBDataFrame}}
}
\keyword{internal}
