% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{coords.data.frame}
\alias{coords.data.frame}
\title{Create a new data.frame with a given coordinate system}
\usage{
\method{coords}{data.frame}(df, new.coords)
}
\arguments{
\item{df}{a data.frame with columns labelled x, y, z (for cartesian)
or theta, phi (for spherical colatitude and longitude respectively)}

\item{new.coords}{specifies the new coordinate system
("spherical" or "cartesian").}
}
\value{
A new data.frame whose coordinates are as specified by
\code{new.coords}
}
\description{
This does not affect the original object \code{df}.
If you wish to affect \code{df} then use
\code{\link{coords<-.data.frame}}
}
\examples{

## Create df with no coords, then create df2 with cartesian coords
df <- data.frame(x = c(1,0,0), y = c(0,1,0), z = c(0,0,1))
df2 <- coords(df, new.coords = "spherical")
df2
df

## Change the coords of df directly (to spherical)
coords(df) <- "spherical"
df

}
