% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{coords}
\alias{coords}
\title{Coordinate conversion generic}
\usage{
coords(x, ...)
}
\arguments{
\item{x}{An object}
}
\description{
Coordinate conversion generic
}
\seealso{
\code{\link{coords.CMBDataFrame}}
\code{\link{coords.CMBWindow}}
\code{\link{coords.HPDataFrame}}
\code{\link{coords.data.frame}}
}
