% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covCMB.R
\name{covCMB}
\alias{covCMB}
\title{Covariance for CMB}
\usage{
covCMB(cmbdf, num.bins = 10, sample.size, max.dist)
}
\arguments{
\item{cmbdf}{is a CMBDataFrame or data.frame}

\item{num.bins}{specifies the number of bins}

\item{sample.size}{optionally specify the size of a simple random
sample to take before calculating covariance. This may be useful if
the full covariance computation is too slow.}

\item{max.dist}{an optional number between 0 and pi specifying the
maximum geodesic distance between any two points in cmbdf.
For example, if cmbdf represents a full sky map
or a random sample of a full sky map then \code{max.dist = pi}.
If max.dist is known then specifying it may reduce
computation time.}
}
\value{

}
\description{
This function provides an empirical covariance estimate for data
in a CMBDataFrame or data.frame. It places data into bins.
}
\examples{

}
