% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CovCMB.R
\name{covCMB}
\alias{covCMB}
\title{Covariance for CMB}
\usage{
covCMB(cmbdf, num.bins = 10, sample.size, max.dist = pi, breaks,
  equiareal = TRUE, calc.max.dist = FALSE)
}
\arguments{
\item{cmbdf}{is a \code{\link{CMBDataFrame}} or \code{data.frame}}

\item{num.bins}{specifies the number of bins}

\item{sample.size}{optionally specify the size of a simple random
sample to take before calculating covariance. This may be useful if
the full covariance computation is too slow.}

\item{max.dist}{an optional number between 0 and pi specifying the
maximum geodesic distance to use for calculating covariance. Only
used if \code{breaks} is unspecified.}

\item{breaks}{optionally specify the breaks manually using a
vector giving the break points between cells. This vector
has length \code{num.bins} since the last break point is taken
as \code{max.dist}. If \code{equiareal = TRUE} then
these breaks should be \eqn{cos(r_i)} where \eqn{r_i} are radii.
If \code{equiareal = FALSE} then these breaks should be \eqn{r_i}.}

\item{equiareal}{if TRUE then the bins have equal spherical
area. If false then the bins have equal annular widths.
Default is TRUE.}

\item{calc.max.dist}{if TRUE then the \code{max.dist} will be
calculated from the locations in cmbdf. Otherwise
either \code{max.dist} must be specified or max.dist will
default to pi.}
}
\value{
An object of class CMBCovariance consisting of a \code{data.frame}
containing sample covariance
values, bin centers, and number \code{n} of data point pairs
whose distance falls in
the corresponding bin.
The first
row of this data.frame corresponds to the sample variance.
The attribute "breaks" contains the break points used.
The returned \code{data.frame} has
\code{num.bins + 1} rows since the first row, the sample
variance, is not counted as a bin.
}
\description{
This function provides an empirical covariance estimate for data
in a CMBDataFrame or data.frame. It places data into bins.
}
\examples{

}
