% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covCMB.R
\name{covPwSp}
\alias{covPwSp}
\title{Covariance estimate via power spectra}
\usage{
covPwSp(PowerSpectra, Ns)
}
\arguments{
\item{PowerSpectra}{a data frame which first column lists values of multipole
moments and the second column gives the corresponding values of CMB power
spectra.}

\item{N}{a number of points in which the covariance estimate is computed on
the interval [-1,1]}
}
\value{
A data frame which first column is 1-d grid starting at -1+1/Ns and
finishing at 1 with the step 2/Ns. The second column is the values of
estimated covariances on this grid.
}
\description{
This function provides a covariance estimate using the values of the estimated
power spectra.
}
\references{
Formula (2.1) in Baran A., Terdik G. Power spectrum estimation
of spherical random fields based on covariances. Annales Mathematicae et
Informaticae 44 (2015) pp. 15–22.
}
