% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DownloadCMBPS.R
\name{DownloadCMBPS}
\alias{DownloadCMBPS}
\title{Download CMB Power Spectra from Planck Legacy Archive.}
\usage{
DownloadCMBPS(link = 1)
}
\arguments{
\item{link}{The URL to download the file}
}
\value{
The CMB Power Spectra Data Frame
}
\description{
The function \code{DownloadCMBPS} download CMB power spectra components from \url{http://pla.esac.esa.int/pla/#cosmology}.
}
\details{
\code{link = 1}: Best-fit LCDM CMB power spectra from the baseline Planck TT,TE,EE+lowE+lensing (2 <= ell <= 2508).

\code{link = 2}: Baseline high-ell Planck TT power spectra (2 <= ell <= 2508).

\code{link = 3}: Baseline high-ell Planck EE power spectra (2 <= ell <= 1996).

\code{link = 4}: Baseline high-ell Planck TE power spectra (2 <= ell <= 1996).

\code{link = 5}: Low-ell Planck EB power spectra (2 <= ell <= 29).

\code{link = 6}: Low-ell Planck BB power spectra (2 <= ell <= 29).
}
\examples{
## Download the Best-fit LCDM CMB power spectra and plot
CMBPS<- DownloadCMBPS(link=1)
plot(CMBPS$L,CMBPS$TT, type="o",col="red",cex=0.3,
     main="CMB Angular Power Spectra",xlab=expression(l),ylab=expression(paste(D[l],"(",mu,K^2,")")))

}
\references{
Planck Legacy Archive \url{http://pla.esac.esa.int/pla/#cosmology}
}
\keyword{CMB}
\keyword{Power}
\keyword{Spectra}
