% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{geoArea.CMBDataFrame}
\alias{geoArea.CMBDataFrame}
\title{Geodesic area covered by a \code{\link{CMBDataFrame}}}
\usage{
\method{geoArea}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame}
}
\value{
the sum of the areas of all pixels (rows) in cmbdf
}
\description{
Gives the surface on the unit sphere
that is encompassed by all pixels in \code{cmbdf}
}
\examples{
## At low resolution, a few data points can
## occupy a large pixel area, e.g.:
cmbdf <- CMBDataFrame(nside = 1, spix = c(1,2,3))
pix(cmbdf)
geoArea(cmbdf) # pi = 1/4*(surface area of unit sphere)
plot(cmbdf, size = 5, hp.boundaries = 1)

}
