% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{geoArea.CMBWindow}
\alias{geoArea.CMBWindow}
\title{Get the geodesic area of a \code{\link{CMBWindow}}}
\usage{
\method{geoArea}{CMBWindow}(win)
}
\arguments{
\item{win}{a CMBWindow}
}
\value{
Tthe spherical area inside win
}
\description{
Get the geodesic area of a \code{\link{CMBWindow}}
}
