% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{geoArea.CMBWindow}
\alias{geoArea.CMBWindow}
\title{Geodesic area of a \code{\link{CMBWindow}}}
\usage{
\method{geoArea}{CMBWindow}(x)
}
\arguments{
\item{x}{A CMBWindow.}
}
\value{
The spherical area inside the CMBWindow \code{x}.
}
\description{
Geodesic area of a \code{\link{CMBWindow}}
}
\examples{

## A window that covers 1/8 of the unit sphere is constructed and its area is
## pi/2 = 1/8*(surface area of unit sphere)

win <- CMBWindow(theta = c(0,pi/2,pi/2), phi = c(0,0,pi/2))
geoArea(win)

}
