% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Geometry.R
\name{geoDist}
\alias{geoDist}
\title{Geodesic distance on the unit sphere}
\usage{
geoDist(p1, p2)
}
\arguments{
\item{p1}{a 3 element vector on the unit sphere
given in cartesian coordinates (x,y,z),
or a 2 element vector (theta, phi) giving spherical coordinates,
can also be a data.frame or matrix with rows specifying vectors}

\item{p2}{a vector on the unit sphere given in cartesian coordinates (x,y,z)
or a named vector (theta, phi) giving spherical coordinates,
can also be a data.frame or matrix with rows specifying vectors}
}
\value{
The geodesic distance between \code{p1} and \code{p2}
}
\description{
Geodesic distance on the unit sphere
}
