% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{header}
\alias{header}
\title{Get the FITS headers from a \code{\link{CMBDataFrame}}}
\usage{
header(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}
}
\value{
The FITS headers belonging to the FITS file from which cmbdf
data was imported
}
\description{
Get the FITS headers from a \code{\link{CMBDataFrame}}
}
\examples{
## First download the map
# downloadCMBMap(foreground = "smica", nside = 1024)
# df <- CMBDataFrame("CMB_map_smica1024.fits")
# df.sample <- CMBDataFrame(df, sample.size = 10000)
# header(df.sample)

}
