% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{HPDataFrame}
\alias{HPDataFrame}
\title{HPDataFrame class}
\usage{
HPDataFrame(..., nside, ordering = "nested", auto.spix = FALSE, spix)
}
\arguments{
\item{...}{data, can be named vectors or a data.frame}

\item{nside}{integer, the nside parameter, i.e, resolution}

\item{ordering}{the HEALPix ordering scheme ("ring" or "nested")}

\item{auto.spix}{boolean. If TRUE then spix will be found from
the coordinates provided in the data. That is, each row of
data will be assigned the pixel index of its closest HEALPix
pixel center. There must be columns x,y,z for cartesian or
theta, phi for spherical colatitude and longitude respectively}

\item{spix}{a vector of HEALPix pixel indices indicating the
pixel locations of the data. Note that \code{spix} is ignored
if \code{auto.spix = TRUE}}
}
\description{
HPDataFrames are a type of \code{data.frame} designed to carry
data located on the unit sphere. Each row of a \code{HPDataFrame}
is associated with a HEALPix pixel index. The \code{HPDataFrame}
also holds an attribute called \code{nside} which stores the
HEALPix Nside parameter (i.e., the resolution of the HEALPix grid
that is being used).
Unlike \code{\link{HPDataFrames}}, HPDataFrames may have
repeated pixel indices. They are made this way so that
multiple data points falling within a given pixel
can be stored in different rows of any given HPDataFrame.
}
