% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDat.R
\name{is.CMBDat}
\alias{is.CMBDat}
\title{Check if an object is of class CMBDat}
\usage{
is.CMBDat(cmbdf)
}
\arguments{
\item{cmbdf}{Any R object}
}
\value{
TRUE if \code{cmbdf} is a CMBDat object, otherwise FALSE
}
\description{
Check if an object is of class CMBDat
}
\examples{
## First download the map
# downloadCMBMap(foreground = "smica", nside = 1024)
# cmbdat <- CMBDat("CMB_map_smica1024.fits", mmap = TRUE)
# class(cmbdat)
# is.CMBDat(cmbdat)

}
