% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{is.CMBWindow}
\alias{is.CMBWindow}
\title{Check if an object is a CMBWindow}
\usage{
is.CMBWindow(win)
}
\arguments{
\item{win}{any object}
}
\value{
TRUE or FALSE depending if win is a CMBWindow
}
\description{
Check if an object is a CMBWindow
}
\examples{

win <- CMBWindow(x=0,y=3/5,z=4/5,r=0.8, set.minus = TRUE)
is.CMBWindow(win)

}
