% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{maxDist.CMBDataFrame}
\alias{maxDist.CMBDataFrame}
\title{Get the maximum distance between all points
in a \code{\link{CMBDataFrame}}}
\usage{
\method{maxDist}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame object}
}
\description{
Get the maximum distance between all points
in a \code{\link{CMBDataFrame}}
}
