% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowGenerics.R
\name{maxDist.CMBWindow}
\alias{maxDist.CMBWindow}
\title{Get the maximum distance between all points
in a \code{\link{CMBWindow}}}
\usage{
\method{maxDist}{CMBWindow}(win)
}
\arguments{
\item{win}{a CMBWindow object}
}
\description{
Get the maximum distance between all points
in a \code{\link{CMBWindow}}
}
