% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{maxDist}
\alias{maxDist}
\title{Get the maximum distance between all points
in a \code{\link{CMBWindow}}}
\usage{
maxDist(win)
}
\arguments{
\item{win}{a CMBWindow object}
}
\description{
Get the maximum distance between all points
in a \code{\link{CMBWindow}}
}
