% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest2ring.R
\name{nest2ring}
\alias{nest2ring}
\title{Nest to Ring.}
\usage{
nest2ring(nSide, Pix)
}
\arguments{
\item{nSide}{is the HEALPix Nside parameter.}

\item{Pix}{is a vector or matrix of HEALPix pixel indices, in the 'nest' ordering scheme.}
}
\value{
the output is a vector or matrix of HEALPix pixel indices in the 'ring' ordering scheme.
}
\description{
\code{nest2ring} converts HEALPix pixel indices in the 'ring' ordering scheme to
HEALPix pixel indices in the 'nest' ordering scheme.
}
\examples{
# compute HEALPix indices in the ring order of the set Pix given in the nest order at Nside
Nside <- 8
Pix <-c(1,2,23)
nest2ring(Nside,Pix)

}
