% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest2ring.R, R/nestRingConversion.R
\name{nest2ring}
\alias{nest2ring}
\alias{nest2ring}
\title{Nest to Ring}
\usage{
nest2ring(CMBdataFrame = cmbdf)

nest2ring(CMBdataFrame = cmbdf)
}
\arguments{
\item{nSide}{is the Nside for HEALPix.}

\item{Pix}{is the set or subset of Pix indices at nSide.}

\item{CMBDataFrame}{A CMB data frame containing a full sky CMB map
in NESTED HEALPix order.}
}
\value{
the output is the corresponding set of Pix in the ring order.

This is a place holder
}
\description{
\code{nest2ring} computes the Pix index in the ring order from the Pix index Pix
in the nest order at nSide.

This is a place holder.
}
