% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestRingConversion.R
\name{nest2ring}
\alias{nest2ring}
\title{Convert from NEST to RING ordering scheme}
\usage{
nest2ring(CMBdataFrame = cmbdf)
}
\arguments{
\item{CMBDataFrame}{A CMB data frame containing a full sky CMB map
in NESTED HEALPix order.}
}
\value{
This is a place holder
}
\description{
This is a place holder.
}
