% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nest_search.R
\name{nest_search}
\alias{nest_search}
\title{nest_search}
\usage{
nest_search(tp, j_2, j_1, pix_1, plot_points)
}
\arguments{
\item{tp}{is the target point on S^2 in spherical coordinates.}

\item{j_2}{is the high level.}

\item{j_1}{is the low level.}

\item{ind_1}{is the initial pix index at level j_1 (in which box to search).}
}
\value{
the output is the list of the Cartesian coordinates and the index
of HEALPix point closest to the target point \code{tp}. That is,
   \code{h$p} is the HEALPix point closest to tp in level j_2 in the subregion
          ind_1 found at level j_1. 
   \code{h$ind} is the pix index of poitn hp at level j_2.
}
\description{
\code{nest_search(tp,j_2,j_1,ind_1)} searches for the HEALPix point closest to tp in level j_2 in the subregion
    found at level j_1. It emploits and requests the nested structure of HEALPix.
    
\code{nest_search(tp,j_2)} searches for the HEALPix point closest to tp in level
    j_2 among all HEALPix points in level j_2.
}
