% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestSearch.R
\name{nestSearch}
\alias{nestSearch}
\title{Nest Search}
\usage{
nestSearch(tp, Nside, index_only, plot_points)
}
\arguments{
\item{tp}{is a row vector of target point in Cartesian coordinates on S^2}

\item{Nside}{is the Nside for which the HEALPix points are searched for}
}
\value{
the output is the HEALPix index only if \code{index_only} TRUE,
the output is the list of the HEALPix index and Cartesian coordinates of
the HEALPix point closest to \code{tp} if \code{index_only} FALSE.
}
\description{
The function \code{nestSearch} computes the nearest HEALPix point at given level 
to a point on S^2.
}
\examples{
# Find the pix index and Cartesian coordinates of the HEALPix point 
# at Nside closest to the target point tp
tp <- c(0,0,1)
h <- nestSearch(tp,Nside=2048,index_only=FALSE,plot_points=TRUE)
cat("Closest HEALPix point to (",tp,") at Nside = 2048 is (",h$p,")")

}
