% Generated by roxygen2: do not edit by hand
<<<<<<< HEAD
% Please edit documentation in R/NestSearch.R
=======
% Please edit documentation in R/nestSearch.R
>>>>>>> 35b1876d6781bb63e95d42038d82f7f11623d604
\name{nestSearch}
\alias{nestSearch}
\title{Nested Search}
\usage{
nestSearch(target, nside, index.only = FALSE, j = 0:log2(nside))
}
\arguments{
\item{target}{A vector of Cartesian coordinates (x,y,z) for the target
point.}

\item{nside}{An integer, the target resolution at which the
resulting pixels are returned.}

\item{j}{A vector of the resolutions to use in each
\code{\link{nestSearch_step}}. This is usually best left as the default.
Changing this parameter is for experienced users only.}
}
\value{
if \code{index.only = TRUE} then the output will be a HEALPix index.
If \code{index.only} FALSE then the output is the list containing the HEALPix index
and Cartesian coordinate vector of the HEALPix point closest to \code{target}
at resolution \code{nside}.
}
\description{
Finds the closest HEALPix pixel center to a given \code{target} point,
specified in Cartesian coordinates, using an efficient nested search
algorithm. HEALPix indices are all assumed to be in the "nested"
ordering scheme.
}
\examples{

}
