% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{nestSearch}
\alias{nestSearch}
\title{Finds the closest pixel center to a point}
\usage{
nestSearch(target, nside, index.only = FALSE, j = 0:log2(nside))
}
\arguments{
\item{target}{A vector of Cartesian coordinates (x,y,z) for the target
point.}

\item{nside}{An integer, the target resolution at which the
resulting pixels are returned.}

\item{j}{A vector of the resolutions to use in each
\code{\link{nestSearch_step}}. This is usually best left as the default.
Changing this parameter is for experienced users only.}
}
\value{
if \code{index.only = TRUE} then the output will be a HEALPix index.
If \code{index.only} FALSE then the output is the list containing the HEALPix index
and Cartesian coordinate vector of the HEALPix point closest to \code{target}
at resolution \code{nside}.
}
\description{
Finds the closest HEALPix pixel center to a given \code{target} point,
specified in Cartesian coordinates, using an efficient nested search
algorithm. HEALPix indices are all assumed to be in the "nested"
ordering scheme.
}
\examples{

## Find the closest HEALPix pixel center at resolution k=1 for
## the point (0.6,0.8,0)

point <- c(0.6,0.8,0)
k <- 1
cpoint <- nestSearch(point, nside=k)

## Plot the closest pixel center in blue and the point (0.6,0.8,0) in red
displayPixels(k,k,plot.j=k, spix=c(cpoint$pix), size=5, incl.labels =FALSE)
rgl::plot3d(point[1], point[2], point[3], col="red", size = 5, add = TRUE)


}
