% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{nestSearch_step}
\alias{nestSearch_step}
\title{nestSearch_step}
\usage{
nestSearch_step(target, j2, pix.j1)
}
\arguments{
\item{target}{is the target point on S^2 in spherical coordinates.}

\item{j2}{is the target resolution.}

\item{pix.j1}{is the initial pix index,
i.e., the j1-level pixel to search in.}
}
\value{
A vector containing the HEALPix pixel index, \code{pix},
of the closest HEALPix pixel center to the target point,
\code{target}, at resolution j2, and its neighbours.
}
\description{
This function is intended for use only with nestSearch.
}
\keyword{internal}
