% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{nside.CMBDataFrame}
\alias{nside.CMBDataFrame}
\title{HEALPix Nside parameter from a CMBDataFrame}
\usage{
\method{nside}{CMBDataFrame}(x)
}
\arguments{
\item{x}{A \code{\link{CMBDataFrame}}.}
}
\value{
The HEALPix Nside parameter.
}
\description{
This function returns the HEALPix Nside parameter of a CMBDataFrame
}
\examples{
df <- CMBDataFrame(nside = 16)
nside(df)

}
