% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{nside}
\alias{nside}
\title{HEALPix Nside parameter from a CMBDataFrame}
\usage{
nside(cmbdf)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}
}
\value{
The HEALPix Nside parameter
}
\description{
This function returns the HEALPix Nside parameter of a CMBDataFrame
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
nside(df)

}
