% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{ordering<-.CMBDataFrame}
\alias{ordering<-.CMBDataFrame}
\title{Assign new ordering scheme to CMBDataFrame}
\usage{
\method{ordering}{CMBDataFrame}(cmbdf, ...) <- value
}
\description{
Assign new ordering scheme to CMBDataFrame
}
\seealso{
\code{\link{ordering.CMBDataFrame}}
}
\keyword{internal}
