% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{ordering}
\alias{ordering}
\title{ordering generic}
\usage{
ordering(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Extra arguments.}
}
\description{
Detailed descriptions and and examples can be found in documentation for specific
ordering functions \code{\link{ordering.CMBDataFrame}}, \code{\link{ordering.HPDataFrame}}
}
\seealso{
\code{\link{ordering.CMBDataFrame}}
\code{\link{ordering.HPDataFrame}}
}
