% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{p2bp}
\alias{p2bp}
\title{Return base pixel to which pixel belongs}
\usage{
p2bp(p, j)
}
\arguments{
\item{p}{The pixel index at resolution j, in nested order.}

\item{j}{The resolution parameter nside = 2^j}
}
\description{
The base pixel to which pixel p belongs at resolution j
}
\examples{

p2bp(5, 0)
p2bp(5, 1)

}
