% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{parent}
\alias{parent}
\title{Return index of parent pixel}
\usage{
parent(p)
}
\arguments{
\item{p}{A pixel index specified in nested order.}
}
\description{
Gives the pixel at resolution \eqn{j - 1} that contains \eqn{p},
where \eqn{p} is specified at resolution \eqn{j} (notice it does
not depend on \eqn{j}).
}
\examples{

 parent(4)
 parent(5)

}
