% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{parent}
\alias{parent}
\title{Return index of parent pixel}
\usage{
parent(p)
}
\arguments{
\item{p}{A pixel index specified in nested order.}
}
\description{
Gives the pixel at resolution j - 1 that contains p,
where p is specified at resoution j (notice it does not depend on j).
}
\examples{

 parent(4)
 parent(5)

}
