% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{pix.HPDataFrame}
\alias{pix.HPDataFrame}
\title{HEALPix pixel indices from \code{\link{HPDataFrame}}}
\usage{
\method{pix}{HPDataFrame}(hpdf, new.pix)
}
\arguments{
\item{hpdf}{a \code{\link{HPDataFrame}}.}

\item{new.pix}{optional vector of pixel indices}
}
\value{
The vector of HEALPix pixel indices (integers) or,
if new.pix is specified,
a new HPDataFrame.
}
\description{
If new.pix is unspecified then this function returns the vector of
HEALPix pixel indices from a HPDataFrame. If new.pix is specified then
this function returns a new HPDataFrame with pixel indices new.pix
}
\examples{
df <- HPDataFrame(I = rep(0,12), nside = 1)
pix(df)

df.new <- pix(df, new.pix= c(1,3,5,10))
pix(df.new)

}
