% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pix2angC}
\alias{pix2angC}
\title{pix2angC}
\arguments{
\item{Nside}{The number of cuts to a HEALPix base resolution pixel.}

\item{Nest}{Set to TRUE for NESTED ordering scheme and FALSE for RING.}

\item{spix}{Optional integer or vector of sample pixel indices.}
}
\description{
Converts HEALPix pixel scheme to spherical coordinates.
}
\details{
This is a place holder
}
