% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pix2vec.R
\name{pix2vec}
\alias{pix2vec}
\title{pix2vec}
\usage{
pix2vec(nSide, order_ring, Pix)
}
\arguments{
\item{nSide}{is the Nside of the HEALPix points.}

\item{order_ring}{TRUE if the HEALPix order is in the ring order; 
FALSE if in the nested order.}

\item{Pix}{is the set of Pix index.}
}
\value{
output is a 3 by length(Pix) matrix of Cartesian coordinates (x,y,z)' of HEALPix points
           at nSide at Pix index in the specified order.
}
\description{
\code{pix2vec} computes the Cartesian cooridinates of the HEALPix points 
   at Pix index at nSide in the specified order.
}
