% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{plot.CMBDataFrame}
\alias{plot.CMBDataFrame}
\title{Plot CMB Data}
\usage{
\method{plot}{CMBDataFrame}(cmbdf, add = FALSE, sample.size, ...)
}
\arguments{
\item{cmbdf}{a CMB Data Frame with either spherical or cartesian coordinates.}

\item{add}{if TRUE then this plot will be added to any existing plot}

\item{sample.size}{optionally specifies the size of a simple random
sample to take before plotting. This can make the plot less
computationally intensive}
}
\value{
A plot of the CMB data
}
\description{
This function produces a plot from a CMB Data Frame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits")
plot(df, sample.size = 800000)

}
