% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{plot.CMBDataFrame}
\alias{plot.CMBDataFrame}
\title{Plot CMB Data}
\usage{
\method{plot}{CMBDataFrame}(cmbdf, add = FALSE, sample.size, type = "p",
  size = 1, box = FALSE, axes = FALSE, aspect = FALSE, col,
  back.col = "black", labels, ...)
}
\arguments{
\item{cmbdf}{a CMB Data Frame with either spherical or cartesian coordinates.}

\item{add}{if TRUE then this plot will be added to any existing plot.
Note that if \code{back.col} (see below) is specified then a new plot
window will be opened and \code{add = TRUE} will have no effect}

\item{sample.size}{optionally specifies the size of a simple random
sample to take before plotting. This can make the plot less
computationally intensive}

\item{type}{a single character indicating the type of item to plot.
Supported types are: 'p' for points, 's' for spheres, 'l' for lines,
'h' for line segments from z = 0, and 'n' for nothing.}

\item{size}{the size of plotted points}

\item{box}{whether to draw a box}

\item{axes}{whether to draw axes}

\item{aspect}{either a logical indicating whether to adjust the
aspect ratio, or a new ratio.}

\item{col}{specify the colour(s) of the plotted points}

\item{back.col}{optionally specifies the background colour of
the plot. This argument is passed to rgl::bg3d.}

\item{labels}{optionally specify a vector of labels to plot,
such as words or vertex indices. If this is specified then
\code{rgl::text3d} is used instead of \code{rgl::plot3d}. Then
\code{length(labels)} must equal \code{nrow(cmbdf)}}

\item{...}{arguments passed to rgl::plot3d}
}
\value{
A plot of the CMB data
}
\description{
This function produces a plot from a CMB Data Frame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits")
plot(df, sample.size = 800000)

}
