% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Statistics.R
\name{plot.variogram}
\alias{plot.variogram}
\title{Plot variogram}
\usage{
\method{plot}{variogram}(x, ...)
}
\arguments{
\item{x}{An object of class variogram.}

\item{...}{Extra arguments passed to plot.default.}
}
\description{
Plot variogram
}
