% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{print.CMBDataFrame}
\alias{print.CMBDataFrame}
\title{Print CMB Data}
\usage{
\method{print}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}
}
\value{
Prints contents of the CMB data frame to the console.
}
\description{
This function neatly prints the contents of a CMB Data Frame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
print(df)
df

}
