% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HPDataFrame.R
\name{print.HPDataFrame}
\alias{print.HPDataFrame}
\title{Print a \code{\link{HPDataFrame}}}
\usage{
\method{print}{HPDataFrame}(x, ...)
}
\arguments{
\item{x}{A HPDataFrame.}

\item{...}{arguments passed to \code{\link{print.tbl_df}}}
}
\value{
Prints contents of the HPDataFrame to the console.
}
\description{
This function neatly prints the contents of a HPDataFrame.
}
\examples{
df <- HPDataFrame(I = rep(0,12), nside = 1, ordering = "nested")
print(df)
df

}
