% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{rbind.CMBDataFrame}
\alias{rbind.CMBDataFrame}
\title{Like \code{\link{rbind}} for CMBDataFrames}
\usage{
\method{rbind}{CMBDataFrame}(..., deparse.level = 1)
}
\description{
Add a new row or rows to a \code{\link{CMBDataFrame}}.
All arguments passed to \code{...} must be CMBDataFrames.
}
\details{
See the documentation for \code{\link{rbind}}
}
