% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcosmo.R
\name{Rcosmo}
\alias{Rcosmo}
\alias{rCosmo}
\title{Handling and Analysing CMB data}
\description{
Handling and Analysing Spherical, Helpix and Cosmic Microwave Background
data on a HEALPix grid.
}
\section{Update}{
 Current updates are available through
URL: https://github.com/VidaliLama/rcosmo
}

\section{BugReports}{
 https://github.com/VidaliLama/rcosmo/issues
}

\section{Depends}{
 R (>= 3.4.0)
}

\section{License}{
 GPL-2
}

\section{Encoding}{
 UTF-8
}

\section{LazyData}{
 true
}

\section{Imports}{

  FITSio (>= 2.1-0),
Rcpp (>= 0.12.11),
mmap (>= 0.6-17),
tibble (>= 1.4.2),
rgl (>= 0.99.16),
cli (>= 1.0.0),
entropy (>= 1.2.1),
geoR (>= 1.7-5.2.1)
}

\section{Suggests}{

  knitr,
rmarkdown,
testthat,
R.rsp,
gsl
}

\section{LinkingTo}{
 Rcpp
}

\author{
Daniel Fryer \email{d.fryer@latrobe.edu.au},
             Andriy Olenko \email{a.olenko@latrobe.edu.au},
             Ming Li,
             Yuguang Wang.
}
\keyword{internal}
