% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcosmo.R
\docType{package}
\name{rcosmo}
\alias{rcosmo}
\alias{rcosmo-package}
\title{rcosmo: A package for analysis of Cosmic Microwave Background data on a sphere}
\description{
The rcosmo package provides twelvety categories of important functions:
foo, bar, baz and others.
}
\section{Foo functions}{

The foo functions...
}

