% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcosmo.R
\docType{package}
\name{rcosmo}
\alias{rcosmo}
\alias{rCosmo}
\alias{Rcosmo}
\alias{rcosmo-package}
\title{rcosmo - This Documentation is a place holder.}
\description{
To be completed
}
\section{Section1}{

To be completed
}

\section{Section2}{

To be completed
}

\section{Section 3}{

To be completed
}

\section{Dependencies}{

To be completed
}

\author{
Daniel Fryer <d.fryer@latrobe.edu.au>
}
