% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{resolution}
\alias{resolution}
\title{Get the arcmin resolution from a \code{\link{CMBDataFrame}}}
\usage{
resolution(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}
}
\value{
The arcmin resolution as specified by the FITS file where the
data was sourced
}
\description{
Get the arcmin resolution from a \code{\link{CMBDataFrame}}
}
\examples{

## First download the map
# downloadCMBMap(foreground = "smica", nside = 1024)
# df <- CMBDataFrame("CMB_map_smica1024.fits")
# resolution(df)

}
