% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conversion.R
\name{ring2nest}
\alias{ring2nest}
\title{Convert from "ring" to "nested" ordering scheme}
\usage{
ring2nest(nside, pix)
}
\arguments{
\item{nside}{is the HEALPix nside parameter}

\item{pix}{is the set or subset of pixel indices at nside}
}
\value{
vector of pixel indices in "nested" ordering scheme
}
\description{
Convert from "ring" to "nested" ordering scheme
}
