% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestRingConversion.R
\name{ring2nest}
\alias{ring2nest}
\title{Convert from RING to NEST ordering scheme}
\usage{
ring2nest(cmbdf)
}
\arguments{
\item{CMBDataFrame}{A CMB data frame containing a full sky CMB map
in RING HEALPix order.}
}
\value{
This is a place holder
}
\description{
This is a place holder.
}
