% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nestSearch.R
\name{siblings}
\alias{siblings}
\title{siblings}
\usage{
siblings(p)
}
\arguments{
\item{p}{Pixel index in nested order.}
}
\description{
The siblings of pixel p are defined as the
children of the parent of p. Note this is resolution independent.
}
