% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NestSearch.R
\name{siblings}
\alias{siblings}
\title{Return siblings of pixel}
\usage{
siblings(p)
}
\arguments{
\item{p}{Pixel index in nested order.}
}
\description{
The siblings of pixel \eqn{p} are defined as the
children of the parent of \eqn{p}. Note this is resolution independent.
}
\examples{

siblings(11)
siblings(12)

}
