% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sph2car}
\alias{sph2car}
\title{sph2car}
\usage{
sph2car(df)
}
\arguments{
\item{df}{a data.frame with columns labelled \code{theta} and \code{phi}
for colatitude and longitude respectively}
}
\value{
a data.frame with columns x, y, z (cartesian coordinates)
}
