% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphericalHarmonics.R
\name{sphericalharmonics}
\alias{sphericalharmonics}
\title{Compute spherical harmonic values at given points on the sphere.}
\usage{
sphericalharmonics(L, m, xyz)
}
\arguments{
\item{L}{The degree of spherical harmonic}

\item{m}{The order number of the degree-L spherical harmonic}

\item{xyz}{Dataframe for given points in 3D cartesian coordinates}
}
\value{
The spherical harmonic values
}
\description{
The function \code{sphericalharmonics} computes the spherical harmonic values on the given 3D Cartesian coordinates.
}
\examples{
## Calculate spherical harmonic value at the point (0,1,0) with L=5, m=2
point<-data.frame(x=0,y=1,z=0)
SH_value<-sphericalharmonics(5,2,point)
## Calculate spherical harmonic values at the point (1,0,0), (0,1,0), (0,0,1) with L=5, m=2
points<-data.frame(diag(3))
SH_values<-sphericalharmonics(5,2,points)

}
\references{
Equation (7) in Hesse, K., Sloan, I. H., & Womersley, R. S. (2010).
Numerical integration on the sphere. In Handbook of Geomathematics (pp. 1185-1219).
Springer Berlin Heidelberg.
}
\keyword{harmonic}
\keyword{spherical}
