% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{subWindow}
\alias{subWindow}
\title{Restrict a \code{\link{CMBDataFrame}} to a \code{\link{CMBWindow}}}
\usage{
subWindow(cmbdf, win)
}
\arguments{
\item{cmbdf}{a \code{\link{CMBDataFrame}}}

\item{win}{a \code{\link{CMBWindow}}}
}
\value{
a \code{\link{CMBDataFrame}} which is restricted to the
region of the sky specified by \code{win}
}
\description{
Restrict a \code{\link{CMBDataFrame}} to a \code{\link{CMBWindow}}
}
\examples{

}
