% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{subWindow}
\alias{subWindow}
\title{subWindow}
\usage{
subWindow(cmbdf, win, intersect = TRUE, in.pixels, in.pixels.res = 0)
}
\arguments{
\item{cmbdf}{a \code{\link{CMBDataFrame}}, a \code{data.frame},
or CMBDat object. If this is a data.frame then it must have
columns labelled x,y,z specifying cartesian coordinates, or
columns labelled theta, phi specifying colatitude and longitude
respectively.}

\item{win}{a \code{\link{CMBWindow}} or a list of CMBWindows}

\item{intersect}{a boolean that determines
the behaviour when \code{win} is a list (see details).}

\item{in.pixels}{a vector of pixels at resolution
\code{in.pixels.res} whose union contains the
window(s) \code{win} entirely. This will only be used
if \code{cmbdf} is a \code{CMBDataFrame}}

\item{in.pixels.res}{a resolution
(i.e., \eqn{j} such that nside = \code{2^j|}) at
which the \code{in.pixels} parameter is specified}
}
\value{
a CMBDataFrame, or just a data.frame,
which is restricted to the
region of the sky specified by \code{win}
}
\description{
Restricts a \code{\link{CMBDataFrame}}, \code{CMBDat} object,
or \code{\link{data.frame}} to a \code{\link{CMBWindow}} region.
A single CMBWindow or a list of CMBWindows can be passed to the \code{win}
argument.
}
\details{
Windows that are tagged with \code{set.minus} (see \code{\link{CMBWindow}})
are treated differently from other windows: Let \eqn{A} be the union of the
interiors of all windows whose winType does not include "minus",
and let \eqn{B} be the intersection of the exteriors of all the windows whose
\code{winType} does include "minus". Then, provided that
\code{intersect = TRUE} (the default), the returned CMBDataFrame will
be the intersection of the points in \code{cmbdf} with \eqn{A} and \eqn{B}.
Otherwise, if \code{intersect = FALSE}, the returned CMBDataFrame will
be the intersection of the points in \code{cmbdf} with the union of
\eqn{A} and \eqn{B}.
Note that if \eqn{A} (resp. \eqn{B}) is empty
then the returned CMBDataFrame will be the intersection of \eqn{B}
(resp. \eqn{A}) with \code{cmbdf}.
}
\examples{

}
