% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{summary.CMBDataFrame}
\alias{summary.CMBDataFrame}
\title{Summarise CMB Data}
\usage{
\method{summary}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMB Data Frame.}
}
\value{
A summary of the CMB data.
}
\description{
This function produces a summary from a CMB Data Frame.
}
\examples{
df <- CMBDataFrame("CMB_map_smica1024.fits", sample.size = 800000)
summary(df)

}
