% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameGenerics.R
\name{summary.CMBDataFrame}
\alias{summary.CMBDataFrame}
\title{Summarise a \code{\link{CMBDataFrame}}}
\usage{
\method{summary}{CMBDataFrame}(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}
}
\value{
A summary
}
\description{
This function produces a summary from a CMBDataFrame.
}
