% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NewGenerics.R
\name{window<-}
\alias{window<-}
\title{window generic}
\usage{
window(x, ...) <- value
}
\description{
Detailed descriptions and and examples can be found in documentation for specific
window functions \code{\link{window.CMBDataFrame}}, \code{\link{window.HPDataFrame}},
\code{\link{window.data.frame}}
}
\seealso{
\code{\link{window.CMBDataFrame}}
\code{\link{window.HPDataFrame}}
\code{\link{window.data.frame}}
}
\keyword{internal}
