% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{window}
\alias{window}
\title{Window attribute of \code{\link{CMBDataFrame}}}
\usage{
window(cmbdf)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}
}
\value{
The window attribute of cmbdf
}
\description{
This function returns the \code{\link{CMBWindow}} attribute of a
CMBDataFrame. The return value is NULL if the window is full sky
}
\examples{
cmbdf <- CMBDataFrame(nside = 16, coords = "cartesian", ordering = "nested")
win <- CMBWindow(theta = c(0,pi/2,pi/2), phi = c(0,0,pi/2))
window(cmbdf) <- win
plot(cmbdf)
window(cmbdf)

}
