% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBDataFrameHelpers.R
\name{window}
\alias{window}
\title{Window attribute of \code{\link{CMBDataFrame}}}
\usage{
window(cmbdf, new.window, intersect = TRUE, in.pixels,
  in.pixels.res = 0)
}
\arguments{
\item{cmbdf}{a CMBDataFrame.}

\item{new.window}{optionally specify a new window
in which case a new CMBDataFrame is returned whose CMBWindow is new.window.
\code{new.window} may also be a list (see details section).}

\item{intersect}{a boolean that determines
the behaviour when \code{win} is a list (see details).}

\item{in.pixels}{a vector of pixels at resolution
\code{in.pixels.res} whose union contains the
window(s) \code{win} entirely, or if \code{new.window} is
unspecified then this whole pixel is returned}

\item{in.pixels.res}{a resolution
(i.e., \eqn{j} such that nside = \code{2^j|}) at
which the \code{in.pixels} parameter is specified}
}
\value{
The window attribute of cmbdf or, if new.window/in.pixels is specified,
a new CMBDataFrame.
}
\description{
When new.window or in.pixels is unspecified this function returns the
\code{\link{CMBWindow}} attribute of a
CMBDataFrame. The return value is NULL if the window is full sky.
When new.window is specified this function instead returns
a new CMBDataFrame whose CMBWindow attribute is new.window
}
\details{
Windows that are tagged with \code{set.minus} (see \code{\link{CMBWindow}})
are treated differently from other windows. See \code{\link{subWindow}} for
more details.
}
\examples{
cmbdf <- CMBDataFrame(nside = 16, coords = "cartesian", ordering = "nested")

## Create a new CMBDataFrame with a window
win <- CMBWindow(theta = c(0,pi/2,pi/2), phi = c(0,0,pi/2))
cmbdf.win <- window(cmbdf, new.window = win)
plot(cmbdf.win)
window(cmbdf.win)

## Change the window of an existing CMBDataFrame
window(cmbdf) <- CMBWindow(theta = rep(0.1, 10),
                           phi = seq(0, 18*pi/10, length.out = 10))
plot(cmbdf)

}
