% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{winType<-}
\alias{winType<-}
\title{Assign new \code{\link{winType}} to a \code{\link{CMBWindow}}}
\usage{
winType(win, ...) <- value
}
\description{
Assign new \code{\link{winType}} to a \code{\link{CMBWindow}}
}
\examples{

win <- CMBWindow(x = 1, y = 0, z = 0, r = 0.5, set.minus = TRUE)
winType(win)
winType(win) <- "disc"
winType(win)

}
\seealso{
\code{\link{winType}}
}
\keyword{internal}
