% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CMBWindowHelpers.R
\name{winType}
\alias{winType}
\title{Get the type (polygon or disk) of a \code{\link{CMBWindow}}}
\usage{
winType(win)
}
\arguments{
\item{win}{a CMBWindow object}
}
\description{
Get the type (polygon or disk) of a \code{\link{CMBWindow}}
}
